### 2024-04-25 #6680 @editorialbot
[PRE REVIEW]: statConfR: An R Package for Static Models of Decision Confidence and Metacognition
**Submitting author:** @ManuelRausch (<a href="http://orcid.org/0000-0002-5805-5544">Manuel Rausch</a>)
**Repository:** https://github.com/ManuelRausch/StatConfR
**Branch with paper.md** (empty if default branch):
**Version:** v.0.1.2
