+++ paper.md
@@ -35,7 +35,7 @@ We present the `statConfR` package for R, which allows researchers to convenient
 
 Cognitive models of confidence are currently used implicitly and explicitly in a wide range of research areas in the cognitive sciences: In perception research, confidence judgments can be used to quantify perceptual sensitivity based on receiver operating characteristics [@egan_operating_1959], a method based on the signal detection rating model [@Green1966; @hautus_detection_2021]. In metacognition research, the most popular measure of metacognitive performance, the meta-d$^\prime$/d$^\prime$ method [@Maniscalco2012; @Maniscalco2014], implicitly relies on the independent truncated Gaussian model [@rausch_measures_2023]. Finally, confidence models have become a flourishing research topic in their own right [@boundy-singer_confidence_2022; @Desender2021; @guggenmos_reverse_2022; @hellmann_confidence_2024; @hellmann_simultaneous_2023; @pereira_evidence_2021; @Rausch2018; @Rausch2020; @Shekhar2020a; @shekhar_how_2023]. However, too few studies have empirically compared different confidence models [@Rausch2018; @Rausch2020; @rausch_measures_2023; @Shekhar2020a; @shekhar_how_2023], so there is still no consensus about the computational principles underlying confidence judgments [@rahnev_consensus_2022]. This is problematic because meta-d$^\prime$/d$^\prime$ can be biased by discrimination sensitivity, discrimination criteria, and/or confidence criteria if the generative model underlying the data is not the independent truncated Gaussian model [@rausch_measures_2023]. Likewise, receiver operating characteristics in rating experiments are only appropriate measures of discrimination sensitivity if the assumptions of the signal detection rating model are correct [@Green1966; @hautus_detection_2021].
 
-At the time of writing, `statConfR` is the only available package for an open software that allows researchers to fit a set of static models of decision confidence. The ReMeta toolbox provides functions for MATLAB to also fit a variety of different confidence models [@guggenmos_reverse_2022], but some important models such as the independent truncated Gaussian model are missing. Previous studies modelling confidence have made their analysis scripts freely available on the OSF website [@rausch_full_2017; @rausch_full_2018; @rausch_full_2022; @shekhar_nature_2020; @shekhar_how_2022], but these analysis scripts are often tailored to specific experiments and require time and effort to adapt to new experiments. In addition, the documentation of these scripts is not always sufficient to be used without export knowledge in cognitive modelling. Finally, the lognormal noise model and the lognormal weighted evidence and visibility model were previously only available implemented in MATLAB, so `statConfR` makes these confidence models available to researchers who do not have access to MATLAB. The `statConfR` package also provides a faithful implementation of meta-d$^\prime$/d$^\prime$, which has been originally implemented in MATLAB [@maniscalco_type_2020]. Fleming provides R code for Hmetad, a Bayesian hierarichical version of meta-d$^\prime$/d$^\prime$ [@fleming_hmeta-d_2017], but notably the model specification used for Hmetad is not the same as in meta-d$^\prime$/d$^\prime$ [@rausch_measures_2023].  
[…]
