+++ paper.bib
@@ -256,7 +256,7 @@
 @article{rahnev_consensus_2022,
 	title = {Consensus Goals in the Field of Visual Metacognition},
 	volume = {17},
-	doi = {10.1177/174569162210756},
+	doi = {10.1177/17456916221075615},
 	abstract = {Despite the tangible progress in psychological and cognitive sciences over the last several years, these disciplines still trail other more mature sciences in identifying the most important questions that need to be solved. Reaching such consensus could lead to greater synergy across different laboratories, faster progress, and increased focus on solving important problems rather than pursuing isolated, niche efforts. Here, 26 researchers from the field of visual metacognition reached consensus on four long-term and two medium-term common goals. We describe the process that we followed, the goals themselves, and our plans for accomplishing these goals. If this effort proves successful within the next few years, such consensus building around common goals could be adopted more widely in psychological science.},
 	pages = {1746--1765},
 	number = {6},
@@ -290,6 +290,7 @@
 @article{egan_operating_1959,
 	title = {Operating Characteristics Determined by Binary Decisions and by Ratings},
 	volume = {31},
+	doi = {10.1121/1.1907783},
 	pages = {768--773},
 	number = {6},
 	journaltitle = {Journal of the Acoustical Society of America},
@@ -326,89 +327,4 @@
 	date = {2024},
 	langid = {english},
 	file = {Hellmann et al. - 2023 - Confidence is influenced by evidence accumulation .pdf:C\:\\Users\\mru\\Zotero\\storage\\XRR256WT\\Hellmann et al. - 2023 - Confidence is influenced by evidence accumulation .pdf:application/pdf},
-}
-
-@online{shekhar_nature_2020,
-	title = {The nature of metacognitive inefficiency in perceptual decision making},
-	url = {https://osf.io/s8fnb/},
-	abstract = {Hosted on the Open Science Framework},
-	author = {Shekhar, Medha and Rahnev, Dobromir},
-	urldate = {2024-04-04},
-	date = {2020-04-03},
-	note = {Publisher: {OSF}},
-	file = {Snapshot:C\:\\Users\\mru\\Zotero\\storage\\DFGNZ5E7\\s8fnb.html:text/html},
-}
-
-@online{shekhar_how_2022,
-	title = {How do humans give confidence? A comprehensive comparison of process models of metacognition},
-	url = {https://osf.io/g8f9x/},
-	shorttitle = {How do humans give confidence?},
-	abstract = {Hosted on the Open Science Framework},
-	author = {Shekhar, Medha and Rahnev, Dobromir},
-	urldate = {2024-04-04},
-	date = {2022-03-30},
-	note = {Publisher: {OSF}},
-	file = {Snapshot:C\:\\Users\\mru\\Zotero\\storage\\7Y9RWYDA\\g8f9x.html:text/html},
-}
-
-@online{rausch_full_2017,
-	title = {Full material to "Confidence in masked orientation judgments is informed by both evidence and visibility"},
-	url = {https://osf.io/ty4h8/},
-	abstract = {Hosted on the Open Science Framework},
-	author = {Rausch, Manuel and Zehetleitner, Michael},
-	urldate = {2024-04-04},
-	date = {2017-04-07},
-	note = {Publisher: {OSF}},
-	file = {Snapshot:C\:\\Users\\mru\\Zotero\\storage\\BCV2C5UX\\ty4h8.html:text/html},
-}
-
-@online{rausch_full_2022,
-	title = {Full material to "Measures of metacognitive efficiency across cognitive models of decision confidence"},
-	url = {https://osf.io/72uds/},
-	abstract = {Abstract
-
[…]
