+++ paper.bib
@@ -76,6 +76,7 @@
 @article{Maniscalco2012,
 	title = {A signal detection theoretic method for estimating metacognitive sensitivity from confidence ratings},
 	volume = {21},
+	doi = {10.1016/j.concog.2011.09.021},
 	pages = {422--430},
 	number = {1},
 	journaltitle = {Consciousness and Cognition},
