### 2024-05-31 #6680 @ManuelRausch
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1016/j.concog.2017.02.007 is OK
> - 10.1038/s41467-021-23540-y is OK
> - 10.3758/s13414-017-1431-5 is OK
> - 10.1038/s41562-019-0813-1 is OK
> - 10.3758/s13414-021-02284-3 is OK
> - 10.1037/a0019737 is OK
> - 10.1093/nc/nix007 is OK
> - 10.1037/rev0000249 is OK
> - 10.7554/eLife.75420 is OK
> - 10.1007/978-3-642-45190-4_3 is OK
> - 10.1093/nc/niw002 is OK
> - 10.1038/s41467-022-31727-0 is OK
> - 10.1016/j.cognition.2020.104522 is OK
> - 10.1016/j.neuroimage.2020.116963 is OK
> - 10.1038/s41562-022-01464-x is OK
> - 10.1037/rev0000411 is OK
> - 10.1037/xge0001524 is OK
> - 10.1037/met0000634 is OK
> - 10.31234/osf.io/5ze8t is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Signal detection theory and psychophysics
> - No DOI given, and none found for title: A signal detection theoretic method for estimating...
> - No DOI given, and none found for title: Detection theory: A user’s guide
> - 10.1121/1.1907783 may be a valid DOI for title: Operating Characteristics Determined by Binary Dec...
> - 10.31234/osf.io/g9qzh may be a valid DOI for title: The nature of metacognitive inefficiency in percep...
> - No DOI given, and none found for title: How do humans give confidence? A comprehensive com...
> - No DOI given, and none found for title: Full material to "Confidence in masked orientation...
> - No DOI given, and none found for title: Full material to "Measures of metacognitive effici...
> - No DOI given, and none found for title: Full material to "Cognitive modelling reveals dist...
> - No DOI given, and none found for title: Type 2 signal detection theory analysis using meta...
> - Entry without DOI or title found
> - No DOI given, and none found for title: HMeta-d
>
> INVALID DOIs
>
> - 10.1177/174569162210756 is INVALID
> ```

I corrected the invalid doi and added various missing dois. The remaining references without doi are books or book sections and to my best knowledge do not have doi.
