### 2024-07-01 #6680 @lwu9
Hi Sam,

After roughly looking through the project, I don't think I'm familiar with
the signal detection in the project. So I'll skip this one. Thanks!

Best,
Lili
