### 2024-07-05 #6680 @samhforbes
Hi @marcoapintoo @christinamaher would either of you be willing to review this package for JOSS?
