### 2024-07-07 #6966 @editorialbot
[REVIEW]: statConfR: An R Package for Static Models of Decision Confidence and Metacognition
**Submitting author:** @ManuelRausch (<a href="http://orcid.org/0000-0002-5805-5544">Manuel Rausch</a>)
**Repository:** https://github.com/ManuelRausch/StatConfR
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.0
**Archive:** 10.17605/OSF.IO/EQ3XK
