### 2024-07-31 #6966 @christinamaher
Hi @samhforbes ! I believe I've gone through the relevant points in the checklist above. The issues (related to points I left unchecked above) are as follows -
**General Checks - Human Research Approvals:** The dataset provided is from a human experiment. Although it appears to have been previously published, the documentation does not provide details regarding informed consent for data sharing or ethical approvals.

**Documentation - Functionality Documentation:** The code would benefit significantly from improved documentation. Specifically:
*Function Documentation:* Each function is lacking explanation of its purpose, parameters, and outputs.
*README File:* The README file should provide a clearer guide to the package's components, including how to use the functions effectively.

**Software Paper - Summary:** The paper lacks a high-level summary of the package’s primary use-case. Including this summary would help readers understand the main objectives and applications of the package.
*Model Descriptions:* Authors should consider providing a detailed description of the models included in the package and an explain for the rationale behind the choice of models and the fitting procedures used.
*Visual Aids:* Incorporate figures or schematics that visualize the models and the results would greatly enhance the paper's readability and package's usability.
*Plotting Utilities:* Consider adding plotting utilities to the package to facilitate visualization of results.
