### 2024-08-05 #6966 @haoxue-fan
Thanks for the opportunity to review the package and sorry for the delayed response! I first want to applaud for the authors @ManuelRausch for their effort putting together this package - it is super useful for researchers in relevant fields both in terms of encouraging them to try out different models as well as lower the coding barrier. I was able to load the R package and run the code as stated in README without difficulty. However, I have a couple of comments listed below most related to the writing and the documentation aspects of the package that I think worth improving:

* Documentation: I share the same feeling as @christinamaher that the package can benefit greatly from more documentation. The way that the paper is currently written and the README file makes me feel that the author assumes a relatively high degree of expertise in the related research domain in the audience, which may or may not be true. Therefore, I think it would be great to provide a step-by-step walkthrough example in the paper. This can be the current example use described in README, but probably with more text descriptions e.g., what the output means; what are some functions that can be used; what are the orders that the functions should be used.

* Relatedly, I notice that on the function help page there are descriptions related to the input data format as well as detailed description of each model. It would be great to make it explicit (e.g., mention that the users can refer to XXX page, or paraphrase them concisely in the paper/README).

* Is it possible for the audience to implement their own models? Is there any guideline on how to do that, and is there any platform for people to share the model with each other? I understand that this maybe too much to ask for, but either way (allowing community contribution or not), it may worth spoiling some ink describing whether the users can implement their customized models, whether this is recommended, and what are some tips that can be shared.

* When I was running the example code, the function `fitConfModels` took a while to run. I am not sure whether this is normal/my computer is too old. Either way, it may worth adding some description on README for the running speed so that the users can calibrate their expectation (I was anxiously checking whether my R is still working since there was no output for a while).
