### 2024-09-21 5570932
Extended the documentation for fitting functions
