### 2024-09-21 fd9de28
More documentation for simulation function; generateDataSDT now also returns a ratings column
