### 2024-09-21 bcd06c3
Add contribution paragraph and hint on fitting duration in  README
