+++ paper.md
@@ -1,47 +1,55 @@
 ---
-title: 'statConfR: An R Package for Static Models of Decision Confidence and Metacognition'
+title: 'statConfR: An R Package for Static Models of Decision Confidence and Measures
+  of Metacognition'
 tags:
-  - Cognitive modelling 
-  - R code
-  - signal detection theory
-  - decision confidence
-  - metacognition
-  - meta-d′/d′
+- Cognitive modelling
+- R code
+- signal detection theory
+- decision confidence
+- metacognition
+- "meta-d′/d′"
+date: "08 October 2024"
+output:
+  pdf_document: default
+  word_document: default
+  html_document:
+    df_print: paged
 authors:
-  - name: Manuel Rausch
-    orcid: 0000-0002-5805-5544
-    equal-contrib: true
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-  - name: Sebastian Hellmann
-    orcid: 0000-0001-6006-5103
-    equal-contrib: true 
-    affiliation: "2, 3"
-affiliations:
- - name: Hochschule Rhein-Waal, Fakultät für Gesellschaft und Ökonomie, Germany
-   index: 1
- - name: Katholische Universität Eichstätt-Ingolstadt, Philosophisch-pädagogische Fakultät, Germany
-   index: 2
- - name: TUM School of Management, Technische Universität München, Germany
-   index: 3
-date: 04 April 2024
+- name: Manuel Rausch
+  orcid: "0000-0002-5805-5544"
+  equal-contrib: true
+  affiliation: 1, 2
+- name: Sebastian Hellmann
+  orcid: "0000-0001-6006-5103"
+  equal-contrib: true
+  affiliation: 2, 3
 bibliography: paper.bib
+affiliations:
+- name: "Hochschule Rhein-Waal, Fakultät Gesellschaft und Ökonomie, Germany"
+  index: 1
+- name: "Katholische Universität Eichstätt-Ingolstadt, Philosophisch-pädagogische
+    Fakultät, Germany"
+  index: 2
+- name: TUM School of Management, Technische Universität München, Germany
+  index: 3
 ---
   
 # Summary
   
-We present the `statConfR` package for R, which allows researchers to conveniently fit and compare nine different static models of decision confidence applicable to binary discrimination tasks with confidence ratings: the signal detection rating model [@Green1966], the Gaussian noise model[@Maniscalco2016], the independent Gaussian model [@Rausch2017], the weighted evidence and visibility model [@Rausch2018], the lognormal noise model [@Shekhar2020a], the lognormal weighted evidence and visibility model [@shekhar_how_2023], the independent truncated Gaussian model [@rausch_measures_2023] based on the model specification used for the original meta-d$^\prime$/d$^\prime$ method [@Maniscalco2012; @Maniscalco2014], and the independent truncated Gaussian model based on the model specification of Hmetad [@Fleming2017a]. In addition, the statConfR package provides functions for estimating meta-d$^\prime$/d$^\prime$, the most widely-used measure of metacognitive efficiency, allowing both @Maniscalco2012's and @Fleming2017a's model specification. 
-Finally, the statConfR package includes an example data set, on which the functions can be directly applied and tested. The data set was previously published in [@hellmann_simultaneous_2023], all participants provided written consent for participating in the experiment and for publishing their anonymized data in scientific repositories. The experimental prodecure was approved by the Ethics Committee of the Katholische Universität Eichstätt-Ingolstadt. 
[…]
