+++ paper.bib
@@ -1,25 +1,173 @@
 
+@article{desender_dynamic_2022,
+	title = {Dynamic influences on static measures of metacognition},
+	volume = {13},
+	doi = {10.1038/s41467-022-31727-0},
+	abstract = {Humans differ in their capability to judge the accuracy of their own choices via confidence judgments. Signal detection theory has been used to quantify the extent to which confidence tracks accuracy via M-ratio, often referred to as metacognitive efficiency. This measure, however, is static in that it does not consider the dynamics of decision making. This could be problematic because humans may shift their level of response caution to alter the tradeoff between speed and accuracy. Such shifts could induce unaccounted-for sources of variation in the assessment of metacognition. Instead, evidence accumulation frameworks consider decision making, including the computation of confidence, as a dynamic process unfolding over time. We draw on evidence accumulation frameworks to examine the influence of response caution on metacognition. Simulation results demonstrate that response caution has an influence on M-ratio. We then tested and confirmed that this was also the case in human participants who were explicitly instructed to either focus on speed or accuracy. We next demonstrated that this association between M-ratio and response caution was also present in an experiment without any reference towards speed. The latter finding was replicated in an independent dataset. In contrast, when data were analyzed with a novel dynamic measure of metacognition, which we refer to as v-ratio, in all of the three studies there was no effect of speed-accuracy tradeoff. These findings have important implications for research on metacognition, such as its measurement, domain-generality, individual differences, and neural correlates.Competing Interest StatementThe authors have declared no competing interest.},
+	number = {1},
+	journal = {Nature Communications},
+	author = {Desender, Kobe and Vermeylen, Luc and Verguts, Tom},
+	year = {2022},
+	pages = {1--30},
+	file = {PDF:C\:\\Users\\PPA714\\Zotero\\storage\\3WET758L\\Desender (2021) Dynamic influences on static measures of metacognition.pdf:application/pdf},
+}
+
+@article{Desender2021,
+	title = {Dynamic expressions of confidence within an evidence accumulation framework},
+	volume = {207},
+	issn = {18737838},
+	url = {https://doi.org/10.1016/j.cognition.2020.104522},
+	doi = {10.1016/j.cognition.2020.104522},
[…]
