+++ paper.md
@@ -47,6 +47,119 @@ At the time of writing, `statConfR` is the only available package for an open so
 
 An important limitation of the models implemented in `statConfR` is that the dynamics of the decision process are not taken into account. This is a problem because confidence judgments are related to the dynamics of decision making [@hellmann_confidence_2024; @Pleskac2010; @Rahnev2020]. However, most previously proposed dynamical models of confidence do not include a parameter to represent metacognitive ability. There is one proposal for a dynamical measure of metacognitive efficiency, the v-ratio [@desender_dynamic_2022], which is based on two-stage signal detection theory [@Pleskac2010], but two-stage signal detection theory has been outperformed by other models in a number of visual discrimination tasks [@hellmann_simultaneous_2023; @hellmann_confidence_2024; @shekhar_how_2024]. Thus, the static confidence models included in `statConfR` may still be useful for many researchers. 
  
+# Step-by-step example
+
+## Installation
+
+The latest released version of the package is available on CRAN via
+```r
+    install.packages("statConfR")
+```
+The easiest way to install the development version is using `devtools`
+and install from GitHub:
+
+    devtools::install_github("ManuelRausch/StatConfR")
+
+## Usage
+
+### Data structure
+
+The package includes a demo data set from a masked orientation
+discrimination task with confidence judgments (Hellmann et al., 2023,
+Exp. 1.
+
+``` r
+library(statConfR)
+data("MaskOri")
+head(MaskOri)
+```
+
+    ##   participant stimulus correct rating diffCond trialNo
+    ## 1           1        0       1      0      8.3       1
+    ## 2           1       90       0      4    133.3       2
+    ## 3           1        0       1      0     33.3       3
+    ## 4           1       90       0      0     16.7       4
+    ## 5           1        0       1      3    133.3       5
+    ## 6           1        0       1      0     16.7       6
+
+Data should be in the form of a data.frame object columns for following
+variables:
+
+- stimulus (factor with 2 levels): The property of the stimulus which
+  defines which response is correct
+- diffCond (factor): The experimental manipulation that is expected to
+  affect discrimination sensitivity
+- correct (0-1): Indicating whether the choice was correct (1) or
+  incorrect(0).
+- rating (factor): A discrete variable encoding the decision confidence
+  (high: very confident; low: less confident)
+- participant (integer): giving the subject ID.
+
+### Fitting
+
+It is strongly recommended that if metacognitive efficiency is to be
+measured using the meta-d′/d′ method that researchers fist determine
+whether the Independent Truncated Gaussian Model, the confidence model
+implied by the meta-d′/d′ method, is an adequate description of the
+data. Using the function fitConfModel, we can fit several confidence
+models to the data of each participant. The argument
+`.parallel=TRUE`allows for parallelization over all but one available
+core.
+
+    fitted_pars <- fitConfModels(MaskOri, models=c("SDT", "WEV"), .parallel = TRUE) 
+
+This parallelizes the fitting process over participant-model
+combinations. The output is then a data frame with one row for each
+participant-model combination and columns for parameters and measures
+for model performance (negative log-likelihood, BIC, AIC and AICc).
+These may be used for quantitative model comparison.
+
[…]
