### 2024-10-09 a9751e4
Add description of models to readme
