+++ README.md
@@ -42,16 +42,106 @@ set of criteria $\theta_{R,i}, i=1,2,...,L-1$, depending on the
 discrimination response $R$ to produce a $L$-step discrete confidence response. The different models
 vary in how $y$ is generated (see below). 
 The parameters shared between all models are:
-- sensitivity parameters $d_1$,...,$d_K$ ($K$: number of difficulty levels)
+- sensitivity parameters $d_1, ..., d_K$ ($K$: number of difficulty levels)
 - decision criterion $c$, 
 - confidence criterion $\theta_{-1,1}, ..., \theta_{-1,L-1}, 
 \theta_{1,1},  ,...,\theta_{1,L-1}$ ($L$: number of confidence categories available for confidence ratings)
 
-### \strong{Signal Detection Rating Model (SDT)}
+### Signal Detection Rating Model (SDT)
 According to SDT, the same sample of sensory evidence is used to generate response and confidence, i.e.,
-$y=x} and the confidence criteria span from the left and
-right side of the decision criterion $c}(Green & Swets, 1966).
+$y=x$. The confidence criteria associated with $R=-1$ are more negative than the 
+decision criterion $c$, whereas the confidence criteria associated with $R=1$ are more positive than $c$(Green & Swets, 1966).
 
+### Gaussian Noise Model (GN)
+According to GN, $y$ is subject to additive noise and assumed to be normally distributed
+around the decision evidence value $x$ with a standard deviation $\sigma$ (Maniscalco & Lau, 2016).
+$\sigma$ is an additional free parameter.
+
+#' ### \strong{Weighted Evidence and Visibility model (WEV)}
+#' WEV assumes that the observer combines evidence about decision-relevant features
+#' of the stimulus with the strength of evidence about choice-irrelevant features
+#' to generate confidence (Rausch et al., 2018). Thus, the WEV model assumes that \eqn{y} is normally
+#' distributed with a mean of \eqn{(1-w)\times x+w \times d_k\times R} and standard deviation \eqn{\sigma}.
+#' The standard deviation quantifies the amount of unsystematic variability
+#' contributing to confidence judgments but not to the discrimination judgments.
+#' The parameter \eqn{w} represents the weight that is put on the choice-irrelevant
+#' features in the confidence judgment. \eqn{w} and \eqn{\sigma} are fitted in
+#' addition to the set of shared parameters.
+#'
+#' ### \strong{Post-decisional accumulation model (PDA)}
+#' PDA represents the idea of on-going information accumulation after the
+#' discrimination choice (Rausch et al., 2018). The parameter \eqn{a} indicates the amount of additional
+#' accumulation. The confidence variable is normally distributed with mean
+#' \eqn{x+S\times d_k\times a} and variance \eqn{a}.
+#' For this model the parameter \eqn{a} is fitted in addition to the shared
+#' parameters.
+#'
+#' ### \strong{Independent Gaussian Model (IG)}
+#' According to IG, \eqn{y} is sampled independently
+#' from \eqn{x} (Rausch & Zehetleitner, 2017). \eqn{y} is normally distributed with a mean of \eqn{a\times d_k} and variance
+#' of 1 (again as it would scale with \eqn{m}). The additional parameter \eqn{m}
+#' represents the amount of information available for confidence judgment
+#' relative to amount of evidence available for the discrimination decision and can
+#'  be smaller as well as greater than 1.
+#'
+#' ### \strong{Independent Truncated Gaussian Model: HMetad-Version (ITGc)}
+#' According to the version of ITG consistent
[…]
