### 2024-10-09 b976a1d
Added model speficicatiion
