+++ README.md
@@ -54,94 +54,94 @@ decision criterion $c$, whereas the confidence criteria associated with $R=1$ ar
 
 ### Gaussian Noise Model (GN)
 According to GN, $y$ is subject to additive noise and assumed to be normally distributed
-around the decision evidence value $x$ with a standard deviation $\sigma$ (Maniscalco & Lau, 2016).
-$\sigma$ is an additional free parameter.
+around the decision evidence value $x$ with a standard deviation $\sigma$, 
+which is an additional free parameter (Maniscalco & Lau, 2016).
 
-#' ### \strong{Weighted Evidence and Visibility model (WEV)}
-#' WEV assumes that the observer combines evidence about decision-relevant features
-#' of the stimulus with the strength of evidence about choice-irrelevant features
-#' to generate confidence (Rausch et al., 2018). Thus, the WEV model assumes that \eqn{y} is normally
-#' distributed with a mean of \eqn{(1-w)\times x+w \times d_k\times R} and standard deviation \eqn{\sigma}.
-#' The standard deviation quantifies the amount of unsystematic variability
-#' contributing to confidence judgments but not to the discrimination judgments.
-#' The parameter \eqn{w} represents the weight that is put on the choice-irrelevant
-#' features in the confidence judgment. \eqn{w} and \eqn{\sigma} are fitted in
-#' addition to the set of shared parameters.
-#'
-#' ### \strong{Post-decisional accumulation model (PDA)}
-#' PDA represents the idea of on-going information accumulation after the
-#' discrimination choice (Rausch et al., 2018). The parameter \eqn{a} indicates the amount of additional
-#' accumulation. The confidence variable is normally distributed with mean
-#' \eqn{x+S\times d_k\times a} and variance \eqn{a}.
-#' For this model the parameter \eqn{a} is fitted in addition to the shared
-#' parameters.
-#'
-#' ### \strong{Independent Gaussian Model (IG)}
-#' According to IG, \eqn{y} is sampled independently
-#' from \eqn{x} (Rausch & Zehetleitner, 2017). \eqn{y} is normally distributed with a mean of \eqn{a\times d_k} and variance
-#' of 1 (again as it would scale with \eqn{m}). The additional parameter \eqn{m}
-#' represents the amount of information available for confidence judgment
-#' relative to amount of evidence available for the discrimination decision and can
-#'  be smaller as well as greater than 1.
-#'
-#' ### \strong{Independent Truncated Gaussian Model: HMetad-Version (ITGc)}
-#' According to the version of ITG consistent
-#' with the HMetad-method (Fleming, 2017; see Rausch et al., 2023), \eqn{y} is sampled independently
-#' from \eqn{x} from a truncated Gaussian distribution with a location parameter
-#' of \eqn{S\times d_k \times m/2} and a scale parameter of 1. The Gaussian distribution of \eqn{y}
-#' is truncated in a way that it is impossible to sample evidence that contradicts
-#' the original decision: If \eqn{R = -1}, the distribution is truncated to the
-#' right of \eqn{c}. If \eqn{R = 1}, the distribution is truncated to the left
-#' of \eqn{c}. The additional parameter \eqn{m} represents metacognitive efficiency,
-#' i.e., the amount of information available for confidence judgments relative to
-#' amount of evidence available for discrimination decisions and  can be smaller
-#' as well as greater than 1.
-#'
-#' ### \strong{Independent Truncated Gaussian Model: Meta-d'-Version (ITGcm)}
-#' According to the version of the ITG consistent
-#' with the original meta-d' method (Maniscalco & Lau, 2012, 2014; see Rausch et al., 2023),
[…]
