+++ README.md
@@ -60,33 +60,31 @@ which is an additional free parameter (Maniscalco & Lau, 2016).
 ### Weighted Evidence and Visibility model (WEV)
 WEV assumes that the observer combines evidence about decision-relevant features 
 of the stimulus with the strength of evidence about choice-irrelevant features 
-to generate confidence (Rausch et al., 2018). Thus, the WEV model assumes that $Y$ is normally 
-distributed with a mean of $(1-w)\times x+w \times d_k\times R} and standard deviation $\sigma$.
+to generate confidence (Rausch et al., 2018). Thus, the WEV model assumes that $y$ is normally 
+distributed with a mean of $(1-w)\times x+w \times d_k\times R$ and standard deviation $\sigma$.
 The standard deviation quantifies the amount of unsystematic variability
 contributing to confidence judgments but not to the discrimination judgments.
 The parameter $w$ represents the weight that is put on the choice-irrelevant
-features in the confidence judgment. $w$ and $\sigma$ are fitted in
+features in the confidence judgment. The parameters $w$ and $\sigma$ are free parameters in
 addition to the set of shared parameters.
 
-### \strong{Post-decisional accumulation model (PDA)}
+### Post-decisional accumulation model (PDA)
 PDA represents the idea of on-going information accumulation after the
 discrimination choice (Rausch et al., 2018). The parameter $a$ indicates the amount of additional
 accumulation. The confidence variable is normally distributed with mean
-$x+S\times d_k\times a$ and variance $a$.
-For this model the parameter $a$ is fitted in addition to the shared
+$x+S\times d_k\times a$ and variance $a$. The parameter $a$ is fitted in addition to the shared
 parameters.
 
-### \strong{Independent Gaussian Model (IG)
-According to IG, $y$ is sampled independently
-from $x$ (Rausch & Zehetleitner, 2017). $y$ is normally distributed with a mean of $a\times d_k$ and variance
-of 1 (again as it would scale with $m$). The additional parameter $m$
-represents the amount of information available for confidence judgment
-relative to amount of evidence available for the discrimination decision and can
- be smaller as well as greater than 1.
+### Independent Gaussian Model (IG)
+According to IG, $y$ is sampled independently from $x$ (Rausch & Zehetleitner, 2017). 
+The variable $y$ is normally distributed with a mean of $a\times d_k$ and variance
+of 1. The additional parameter $m$ represents the amount of information available for confidence judgment
+relative to amount of evidence available for the discrimination decision and can 
+be smaller as well as greater than 1.
 
-### \strong{Independent Truncated Gaussian Model: HMetad-Version (ITGc)
-According to the version of ITG consistent
-with the HMetad-method (Fleming, 2017; see Rausch et al., 2023), $y$ is sampled independently
+### Independent Truncated Gaussian Model: HMetad-Version (ITGc)
+According to the version of ITG consistent with the HMetad-method 
+(Fleming, 2017; see Rausch et al., 2023), $y$ is sampled independently
 from $x$ from a truncated Gaussian distribution with a location parameter
 of $S\times d_k \times m/2$ and a scale parameter of 1. The Gaussian distribution of $y$
 is truncated in a way that it is impossible to sample evidence that contradicts
@@ -97,28 +95,26 @@ i.e., the amount of information available for confidence judgments relative to
 amount of evidence available for discrimination decisions and  can be smaller
 as well as greater than 1.
 
-### \strong{Independent Truncated Gaussian Model: Meta-d'-Version (ITGcm)
-According to the version of the ITG consistent
[…]
