+++ R/fitConf.R
@@ -1,13 +1,13 @@
 #' Fit a static confidence model to data
 #'
-#' This function fits the parameters of one static model of decision confidence,
+#' The `fitConf` function fits the parameters of one static model of decision confidence,
 #' provided by the `model` argument, to binary choices and confidence judgments.
 #' See Details for the mathematical specification of the implemented models and
 #' their parameters.
 #' Parameters are fitted using a maximum likelihood estimation method with a
-#' initial grid search to find promissing starting values for the optimization.
-#' Additional measures for model fit (negative log-likelihood, BIC, AIC, and AICc)
-#' are computed, which may be used for a quantiative model evaluation.
+#' initial grid search to find promising starting values for the optimization.
+#' In addition, several measures of model fit (negative log-likelihood, BIC, AIC, and AICc)
+#' are computed, which can be used for a quantitative model evaluation.
 #'
 #' @param data  a `data.frame` where each row is one trial, containing following
 #' variables:
@@ -69,19 +69,19 @@
 #' How the confidence variable \eqn{y} is computed varies across the different models.
 #' The following models have been implemented so far:
 #'
-#' ### \strong{Signal Detection Rating Model (SDT)}
+#' ### \strong{Signal detection rating model (SDT)}
 #' According to SDT, the same sample of sensory
 #' evidence is used to generate response and confidence, i.e.,
 #' \eqn{y=x} and the confidence criteria span from the left and
 #' right side of the decision criterion \eqn{c} (Green & Swets, 1966).
 #'
-#' ### \strong{Gaussian Noise Model (GN)}
+#' ### \strong{Gaussian noise model (GN)}
 #' According to the model, \eqn{y} is subject to
 #' additive noise and assumed to be normally distributed around the decision
 #' evidence value \eqn{x} with a standard deviation \eqn{\sigma} (Maniscalco & Lau, 2016).
 #' The parameter  \eqn{\sigma} is a free parameter.
 #'
-#' ### \strong{Weighted Evidence and Visibility model (WEV)}
+#' ### \strong{Weighted evidence and visibility model (WEV)}
 #' WEV assumes that the observer combines evidence about decision-relevant features
 #' of the stimulus with the strength of evidence about choice-irrelevant features
 #' to generate confidence (Rausch et al., 2018). Here, we use the version of the WEV model
@@ -101,7 +101,7 @@
 #' For this model the parameter \eqn{b} is fitted in
 #' addition to the set of shared parameters.
 #'
-#' ### \strong{Independent Gaussian Model (IG)}
+#' ### \strong{Independent Gaussian model (IG)}
 #' According to IG, \eqn{y} is sampled independently
 #' from \eqn{x} (Rausch & Zehetleitner, 2017). \eqn{y} is normally distributed with a mean of \eqn{a\times d_k} and variance
 #' of 1 (again as it would scale with \eqn{m}). The free parameter \eqn{m}
@@ -109,7 +109,7 @@
 #' relative to amount of evidence available for the discrimination decision and can
 #'  be smaller as well as greater than 1.
 #'
-#' ### \strong{Independent Truncated Gaussian Model: HMetad-Version (ITGc)}
+#' ### \strong{Independent truncated Gaussian model: HMetad-Version (ITGc)}
 #' According to the version of ITG consistent
 #' with the HMetad-method (Fleming, 2017; see Rausch et al., 2023), \eqn{y} is sampled independently
 #' from \eqn{x} from a truncated Gaussian distribution with a location parameter
@@ -122,7 +122,7 @@
[…]
