+++ R/fitConfModels.R
@@ -1,6 +1,6 @@
 #' Fit several static confidence models to multiple participants
 #'
-#' This functions fits the parameters of several computational models of decision
+#' The `fitConfModels` function fits the parameters of several computational models of decision
 #' confidence, in binary choice tasks,  specified in the `model` argument, to
 #' different subsets of one data frame, indicated by different values in the column
 #' `participant` of the `data` argument.
@@ -9,8 +9,10 @@
 #' of model in the `models` argument and sub-data frame of `data` for each value
 #' in the `participant` column.
 #' See Details for more information about the parameters.
-#' In addition to the estimated parameters, measures of model fits are provided
-#' for a quantiative comparison between the models.
+#' Parameters are fitted using a maximum likelihood estimation method with a
+#' initial grid search to find promising starting values for the optimization.
+#' In addition, several measures of model fit (negative log-likelihood, BIC, AIC, and AICc)
+#' are computed, which can be used for a quantitative model evaluation.
 #'
 #' @param data  a `data.frame` where each row is one trial, containing following
 #' variables:
@@ -98,19 +100,19 @@
 #' How the confidence variable \eqn{y} is computed varies across the different models.
 #' The following models have been implemented so far:
 #'
-#' ### \strong{Signal Detection Rating Model (SDT)}
+#' ### \strong{Signal detection rating model (SDT)}
 #' According to SDT, the same sample of sensory
 #' evidence is used to generate response and confidence, i.e.,
 #' \eqn{y=x} and the confidence criteria span from the left and
 #' right side of the decision criterion \eqn{c}(Green & Swets, 1966).
 #'
-#' ### \strong{Gaussian Noise Model (GN)}
+#' ### \strong{Gaussian noise model (GN)}
 #' According to the model, \eqn{y} is subject to
 #' additive noise and assumed to be normally distributed around the decision
 #' evidence value \eqn{x} with a standard deviation \eqn{\sigma}(Maniscalco & Lau, 2016).
 #' \eqn{\sigma} is an additional free parameter.
 #'
-#' ### \strong{Weighted Evidence and Visibility model (WEV)}
+#' ### \strong{Weighted evidence and visibility model (WEV)}
 #' WEV assumes that the observer combines evidence about decision-relevant features
 #' of the stimulus with the strength of evidence about choice-irrelevant features
 #' to generate confidence (Rausch et al., 2018). Thus, the WEV model assumes that \eqn{y} is normally
@@ -129,7 +131,7 @@
 #' For this model the parameter \eqn{a} is fitted in addition to the shared
 #' parameters.
 #'
-#' ### \strong{Independent Gaussian Model (IG)}
+#' ### \strong{Independent Gaussian model (IG)}
 #' According to IG, \eqn{y} is sampled independently
 #' from \eqn{x} (Rausch & Zehetleitner, 2017). \eqn{y} is normally distributed with a mean of \eqn{a\times d_k} and variance
 #' of 1 (again as it would scale with \eqn{m}). The additional parameter \eqn{m}
@@ -137,7 +139,7 @@
 #' relative to amount of evidence available for the discrimination decision and can
 #'  be smaller as well as greater than 1.
 #'
-#' ### \strong{Independent Truncated Gaussian Model: HMetad-Version (ITGc)}
+#' ### \strong{Independent truncated Gaussian model: HMetad-Version (ITGc)}
 #' According to the version of ITG consistent
[…]
