+++ DESCRIPTION
@@ -31,4 +31,4 @@ NeedsCompilation: no
 Packaged: 2023-09-10
 Repository: CRAN
 Roxygen: list(old_usage = TRUE, markdown=TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.2

[…]
