+++ README.md
@@ -1,12 +1,11 @@
-# statConfR: Models of Decision Confidence and Measures of Metacognition
-
 The `statConfR` package provides functions to fit static models of
 decision-making and confidence derived from signal detection theory for
-binary discrimination tasks, as well as meta-d′/d′ as measures of metacognition.
-Fitting models of confidence can be used to test the assumptions underlying
-meta-d′/d′. Several models include a metacognition parameter that may
-serve as an alternative when the assumptions of meta-d′/d′ assuming the
-corresponding model provides a better fit to the data. The following models are included:
+binary discrimination tasks, as well as meta-d′/d′ as measures of
+metacognition. Fitting models of confidence can be used to test the
+assumptions underlying meta-d′/d′. Several models include a
+metacognition parameter that may serve as an alternative when the
+assumptions of meta-d′/d′ assuming the corresponding model provides a
+better fit to the data. The following models are included:
 
 - signal detection rating model,
 
@@ -26,135 +25,172 @@ corresponding model provides a better fit to the data. The following models are 
 - lognormal weighted evidence and visibility model.
 
 ## Mathematical description of implemented models of confidence
-The models included in the statConfR package are all based on signal detection theory (Green & Swets, 1966). 
-It is assumed that participants select a binary discrimination response $R$ about a stimulus $S$.
-Both $S$ and $R$ can be either -1 or 1. $R$ is considered correct if $S=R$.
-In addition, we assume that in the experiment, there are $K$ different levels of stimulus discriminability, 
-i.e. a physical variable that makes the discrimination task easier or harder.
-For each level of discriminability, the function fits a different discrimination
-sensitivity parameter $d_k$. If there is more than one sensitivity parameter,
-we assume that the sensitivity parameters are ordered such as $0 < d_1 < d_2 < ... < d_K$.
-The models assume that the stimulus generates normally distributed sensory evidence $x$ with mean $S\times d_k/2$
-and variance of 1. The sensory evidence $x$ is compared to a decision criterion $c$ 
-to generate a discrimination response $R$, which is 1, if $x$ exceeds $c$ and -1 else.
-To generate confidence, it is assumed that the confidence variable $y$ is compared to another
-set of criteria $\theta_{R,i}, i=1,2,...,L-1$, depending on the
-discrimination response $R$ to produce a $L$-step discrete confidence response. The different models
-vary in how $y$ is generated (see below). 
-The parameters shared between all models are:
-- sensitivity parameters $d_1, ..., d_K$ ($K$: number of difficulty levels),
-- decision criterion $c$, 
-- confidence criterion $\theta_{-1,1}, ..., \theta_{-1,L-1}, 
-\theta_{1,1},  ,...,\theta_{1,L-1}$ ($L$: number of confidence categories available for confidence ratings).
+
+The models included in the statConfR package are all based on signal
+detection theory (Green & Swets, 1966). It is assumed that participants
+select a binary discrimination response $`R`$ about a stimulus $`S`$.
+Both $`S`$ and $`R`$ can be either -1 or 1. $`R`$ is considered correct
+if $`S=R`$. In addition, we assume that in the experiment, there are
+$`K`$ different levels of stimulus discriminability, i.e. a physical
+variable that makes the discrimination task easier or harder. For each
+level of discriminability, the function fits a different discrimination
+sensitivity parameter $`d_k`$. If there is more than one sensitivity
+parameter, we assume that the sensitivity parameters are ordered such as
[…]
