+++ README.rmd
@@ -162,14 +162,8 @@ under the assumption that participants had perfect access to the sensory evidenc
 and were perfectly consistent in placing their confidence criteria (Maniscalco & Lau, 2012, 2014).
 Using a signal detection model describing the primary task to quantify metacognition
 allows a direct comparison between metacognitive accuracy and discrimination performance
-because both are measured on the same scale. Meta-d′ can be compared against the estimate of the distance between the two stimulus distributions 
-estimated from discrimination responses, which is referred to as d′: 
-If meta-d′ equals d′, it means that metacognitive accuracy is exactly as good as expected from discrimination performance. If
-meta-d′ is lower than d′, it means that metacognitive accuracy is suboptimal. It can be shown that 
-the implicit model of confidence underlying the meta-d'/d' method is identical to the independent truncated Gaussian model (Rausch et al., 2023). 
-We strongly recommend that if metacognitive efficiency is to be
-measured using the meta-d′/d′ method that researchers fist determine
-whether the independent truncated Gaussian model, is an adequate description of the
+because both are measured on the same scale. Meta-d′ can be compared against the estimate of the distance between the two stimulus distributions estimated from discrimination responses, which is referred to as d′: If meta-d′ equals d′, it means that metacognitive accuracy is exactly as good as expected from discrimination performance. If meta-d′ is lower than d′, it means that metacognitive accuracy is not optimal. It can be shown that the implicit model of confidence underlying the meta-d'/d' method is identical to different versions of the independent truncated Gaussian model (Rausch et al., 2023), depending on whether the original model specification by Maniscalco and Lau (2012) or alternatively the specification by Fleming (2017). We strongly recommend that if metacognitive efficiency is to bemeasured using the meta-d′/d′ method that researchers fist determine
+whether the independent truncated Gaussian models are adequate descriptions of the
 data. 
 
 ## Installation
@@ -189,11 +183,11 @@ devtools::install_github("ManuelRausch/StatConfR")
 
 ## Usage
 
-### Data structure
+### Example data set
 
 The package includes a demo data set from a masked orientation
 discrimination task with confidence judgments (Hellmann et al., 2023,
-Exp. 1.
+Exp. 1).
 
 ```{r}
 library(statConfR)
@@ -201,9 +195,12 @@ data("MaskOri")
 head(MaskOri)
 ```
 
-Data should be in the form of a data.frame object columns for following
-variables:
+### Fitting
 
+The function `fitConfModels` allows the user to fit several confidence
+models separately to the data of each participant. 
+The data should be provided via the argument `.data` in the form of a data.frame 
+object with the following variables in seperate columns: 
 -   stimulus (factor with 2 levels): The property of the stimulus which
[…]
