+++ README.rmd
@@ -1,7 +1,7 @@
 ---
 title: "README"
 author: "Sebastian Hellmann and Manuel Rausch"
-date: "2024-04-10"
+date: "2024-10-13"
 output:
   md_document:
     variant: gfm
@@ -10,7 +10,7 @@ editor_options:
     wrap: 72
 ---
 
-```{r setup, include=FALSE}
+```{r setup, include=FALSE} # this stuff is invissible in the rendered document
 knitr::opts_chunk$set(echo = TRUE)
 gitrep <- "https://github.com/ManuelRausch/StatConfR/tree"
 gitbranch <- "main/"
@@ -180,6 +180,7 @@ and install from GitHub:
 ```         
 devtools::install_github("ManuelRausch/StatConfR")
 ```
+<!-- without any dots, the code chunk will be shown, but not executed -->
 
 ## Usage
 
@@ -214,9 +215,9 @@ The argument `model` is used to specify which model should be fitted, with 'WEV'
 
 Setting the optional argument `.parallel=TRUE` parallizes model fitting over all but 1 available core.
 Note that the fitting procedure takes may take a considerable amount of time, 
-especially, when there are multiple models, several difficulty conditions, and/or  several confidence categories.  Depending on the model and the number of experimental conditions, fitting one participant and model combination may take up to 20-30 minutes on a modern 2.8GHz CPU.
+especially, when there are multiple models, several difficulty conditions, and/or  several confidence categories.  Depending on the model and the number of experimental conditions, fitting ther WEV model to one participant may take 20-30 minutes on a modern 2.8GHz CPU.
 
-```     
+```{r}     
 fitted_pars <- fitConfModels(MaskOri, models=c("ITGcm", "WEV"), .parallel = TRUE) 
 ```
 
@@ -230,58 +231,97 @@ head(fitted_pars)
 
 It can be seen that the independent truncated Gaussian model is consistently outperformed by the weighted evidence and visibility model, which is why we would not recommend using meta-d′/d′ for this specific task. 
 
+### Visualization 
+
+After obtaining model fits, it is strongly recommended to visualize the prediction implied by the best fitting sets of parameters and to compare the prediction with the actual data. The best way to visualize the data is highly specific to the data set and research question, which is why `statConfR` does not come with its own visualisation tools. This being said, here is an example for how a visualization could look like: 
+
+<!-- insert a visualization  -->
+
+```{r, echo=TRUE, fig.show='hold'}
+library(tidyverse)
+
+AggregatedData <- MaskOri %>%
+  mutate(ratings = as.numeric(rating), diffCond = as.numeric(diffCond)) %>%
+  group_by(participant, diffCond, correct ) %>% 
+  dplyr::summarise(ratings=mean(ratings,na.rm=T)) %>%
+  Rmisc::summarySEwithin(measurevar = "ratings",
+                         withinvars = c("diffCond", "correct"), 
+                         idvar = "participant",
+                         na.rm = TRUE, .drop = TRUE) %>% 
+  mutate(diffCond = as.numeric(diffCond))
+
+AggregatedPrediction <- 
+  rbind(fitted_pars %>%
+          filter(model=="ITGcm") %>%
+          group_by(participant) %>%
+          simConf(model="ITGcm") %>% 
+          mutate(model="ITGcm"), 
+        fitted_pars %>%
+          filter(model=="WEV") %>%
+          group_by(participant) %>%
+          simConf(model="WEV") %>% 
+          mutate(model="WEV")) %>%
+  mutate(ratings = as.numeric(rating) ) %>%
+  group_by(participant, diffCond, correct, model ) %>% 
[…]
