+++ README.md
@@ -1,3 +1,5 @@
+`{r setup, include=FALSE} # this stuff is invissible in the rendered document knitr::opts_chunk$set(echo = TRUE) gitrep <- "https://github.com/ManuelRausch/StatConfR/tree" gitbranch <- "main/"`
+
 The `statConfR` package provides functions to fit static models of
 decision-making and confidence derived from signal detection theory for
 binary discrimination tasks, as well as meta-d′/d′ as measures of
@@ -184,13 +186,15 @@ two stimulus distributions estimated from discrimination responses,
 which is referred to as d′: If meta-d′ equals d′, it means that
 metacognitive accuracy is exactly as good as expected from
 discrimination performance. If meta-d′ is lower than d′, it means that
-metacognitive accuracy is suboptimal. It can be shown that the implicit
-model of confidence underlying the meta-d’/d’ method is identical to the
-independent truncated Gaussian model (Rausch et al., 2023). We strongly
-recommend that if metacognitive efficiency is to be measured using the
-meta-d′/d′ method that researchers fist determine whether the
-independent truncated Gaussian model, is an adequate description of the
-data.
+metacognitive accuracy is not optimal. It can be shown that the implicit
+model of confidence underlying the meta-d’/d’ method is identical to
+different versions of the independent truncated Gaussian model (Rausch
+et al., 2023), depending on whether the original model specification by
+Maniscalco and Lau (2012) or alternatively the specification by Fleming
+(2017). We strongly recommend that if metacognitive efficiency is to
+bemeasured using the meta-d′/d′ method that researchers fist determine
+whether the independent truncated Gaussian models are adequate
+descriptions of the data.
 
 ## Installation
 
@@ -203,13 +207,15 @@ and install from GitHub:
 
     devtools::install_github("ManuelRausch/StatConfR")
 
+<!-- without any dots, the code chunk will be shown, but not executed -->
+
 ## Usage
 
-### Data structure
+### Example data set
 
 The package includes a demo data set from a masked orientation
 discrimination task with confidence judgments (Hellmann et al., 2023,
-Exp. 1.
+Exp. 1).
 
 ``` r
 library(statConfR)
@@ -225,144 +231,259 @@ head(MaskOri)
     ## 5           1        0       1      3    133.3       5
     ## 6           1        0       1      0     16.7       6
 
-Data should be in the form of a data.frame object columns for following
-variables:
-
-- stimulus (factor with 2 levels): The property of the stimulus which
-  defines which response is correct
-- diffCond (factor): The experimental manipulation that is expected to
-  affect discrimination sensitivity
-- correct (0-1): Indicating whether the choice was correct (1) or
-  incorrect(0).
-- rating (factor): A discrete variable encoding the decision confidence
-  (high: very confident; low: less confident)
-- participant (integer): giving the subject ID.
-
 ### Fitting
 
-It is strongly recommended that if metacognitive efficiency is to be
-measured using the meta-d′/d′ method that researchers fist determine
-whether the Independent Truncated Gaussian Model, the confidence model
-implied by the meta-d′/d′ method, is an adequate description of the
-data. Using the function fitConfModels, we can fit several confidence
-models to the data of each participant.
+The function `fitConfModels` allows the user to fit several confidence
[…]
