+++ README.md
@@ -1,5 +1,3 @@
-`{r setup, include=FALSE} # this stuff is invissible in the rendered document knitr::opts_chunk$set(echo = TRUE) gitrep <- "https://github.com/ManuelRausch/StatConfR/tree" gitbranch <- "main/"`
-
 The `statConfR` package provides functions to fit static models of
 decision-making and confidence derived from signal detection theory for
 binary discrimination tasks, as well as meta-d′/d′ as measures of
@@ -397,7 +395,12 @@ PlotMeans <-
 PlotMeans 
 ```
 
-![](README_files/figure-gfm/unnamed-chunk-4-1.png)<!-- -->
+<figure>
+<img src="README_files/figure-gfm/unnamed-chunk-4-1.png"
+alt="Predicted vs observed confidence as a function of discriminability and correctness" />
+<figcaption aria-hidden="true">Predicted vs observed confidence as a
+function of discriminability and correctness</figcaption>
+</figure>
 
 ### Measuring metacognition
 
@@ -424,28 +427,35 @@ implement in in the package.
 ### Instruction for implementing custom models of decision confidence (only recommended for users experienced with cognitive modelling)
 
 For readers who want to use our open code to implement models of
-confidence themselves, the following steps need to be taken: -Derive the
-likelihood of a binary response ($`R=-1, 1`$) and a specific level of
-confidence ($`C=1,...K`$) according to the custom model and a set of
-parameters, given the binary stimulus ($`S=-1, 1`$),
-i.e. $`P(R, C | S)`$. - Use one of the files named ‘int_ll*model*.R’
-from the package sources and adapt the likelihood function according to
-your model. According to our convention, name the new file a
-‘int_ll*yourmodelname*.R’. Note that all parameters are fitted on the
-reals, i.e. positive parameters should be transformed outside the
-log-likelihood function (e.g. using the logarithm) and back-transformed
-within the log-likelihood function (e.g. using the exponential). - Use
-one of the files ‘int_fit*model*.R’ from the package sources and adapt
-the fitting function to reflect the new model. - the initial grid used
-during the grid search should include a plausible range of all
-parameters of your model. - if applicable, the parameters of the initial
-grid needs be transformed so the parameter vector for optimization is
-real-valued) - the optimization routine should call the new
-log-likelihood function. - if applicable, the parameter vector
-i`obtained during optimization needs to be back-transformation for the the output object`res`- Name the new file according to the convention 'int_fit*yourmodelname*.R' - Add your model and fitting-functions to the high-level functions`fitConf`and`fitConfModels\` -
-Add a simulation function in the file ‘int_simulateConf.R’ which uses
-the same structure as the other functions but adapt the likelihood of
-the responses.
+confidence themselves, the following steps need to be taken:
+
+- Derive the likelihood of a binary response ($`R=-1, 1`$) and a
+  specific level of confidence ($`C=1,...K`$) according to the custom
+  model and a set of parameters, given the binary stimulus
+  ($`S=-1, 1`$), i.e. $`P(R, C | S)`$.
+- Use one of the files named ‘int_ll*model*.R’ from the package sources
+  and adapt the likelihood function according to your model. According
+  to our convention, name the new file a ‘int_ll*yourmodelname*.R’. Note
+  that all parameters are fitted on the reals, i.e. positive parameters
[…]
