+++ README.rmd
@@ -10,11 +10,13 @@ editor_options:
     wrap: 72
 ---
 
-```{r setup, include=FALSE} # this stuff is invissible in the rendered document
+
+```{r setup, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE)
 gitrep <- "https://github.com/ManuelRausch/StatConfR/tree"
 gitbranch <- "main/"
 ```
+
 
 The `statConfR` package provides functions to fit static models of
 decision-making and confidence derived from signal detection theory for
@@ -237,7 +239,9 @@ After obtaining model fits, it is strongly recommended to visualize the predicti
 
 <!-- insert a visualization  -->
 
-```{r, echo=TRUE, fig.show='hold'}
+
+```{r, echo=TRUE, results="hide", fig.cap = "Predicted vs observed confidence as a function of discriminability and correctness"}
+
 library(tidyverse)
 
 AggregatedData <- MaskOri %>%
@@ -299,20 +303,17 @@ MetaDs <- fitMetaDprime(data = MaskOri, model="ML", .parallel = TRUE)
 The package is under active development. We are planning to implement new models of decision confidence when they are published. Please feel free to [contact us](malto::manuel.rausch@ku.de) to suggest models to implement in in the package. 
 
 ### Instruction for implementing custom models of decision confidence (only recommended for users experienced with cognitive modelling)
-For readers who want to use our open code to implement models of confidence themselves, 
-the following steps need to be taken: 
--Derive the likelihood of a binary response ($R=-1, 1$) and a specific level of confidence ($C=1,...K$) according to the custom model and a set of parameters, given the binary stimulus ($S=-1, 1$), i.e. $P(R, C | S)$. 
-- Use one of the files named 'int_ll*model*.R' from the package sources and adapt the likelihood function according to your model. According to our convention, name the new file a 'int_ll*yourmodelname*.R'. Note that all parameters are fitted on the reals, i.e. positive parameters 
-  should be transformed outside the log-likelihood function (e.g. using the 
-  logarithm) and back-transformed within the log-likelihood function (e.g. using
-  the exponential).
+For readers who want to use our open code to implement models of confidence themselves, the following steps need to be taken: 
+
+- Derive the likelihood of a binary response ($R=-1, 1$) and a specific level of confidence ($C=1,...K$) according to the custom model and a set of parameters, given the binary stimulus ($S=-1, 1$), i.e. $P(R, C | S)$. 
+- Use one of the files named 'int_ll*model*.R' from the package sources and adapt the likelihood function according to your model. According to our convention, name the new file a 'int_ll*yourmodelname*.R'. Note that all parameters are fitted on the reals, i.e. positive parameters should be transformed outside the log-likelihood function (e.g. using the logarithm) and back-transformed within the log-likelihood function (e.g. usingthe exponential).
 - Use one of the files 'int_fit*model*.R'  from the package sources and adapt the fitting function to reflect the new model. 
-  -   the initial grid used during the grid search should include a plausible range of all  parameters of your model. 
-  -   if applicable, the parameters of the initial grid needs be transformed so the
++ the initial grid used during the grid search should include a plausible range of all  parameters of your model. 
++ if applicable, the parameters of the initial grid needs be transformed so the
       parameter vector for optimization is real-valued)
[…]
