+++ paper.md
@@ -8,7 +8,7 @@ tags:
 - decision confidence
 - metacognition
 - "meta-d′/d′"
-date: "08 October 2024"
+date: "12 October 2024"
 output:
   pdf_document: default
   word_document: default
@@ -38,6 +38,7 @@ affiliations:
   
 We present the `statConfR` package for R, which allows researchers to conveniently fit and compare nine different static models of decision confidence applicable to binary discrimination tasks with confidence ratings: the signal detection rating model [@Green1966], the Gaussian noise model [@Maniscalco2016], the independent Gaussian model [@Rausch2017], the weighted evidence and visibility model [@Rausch2018], the lognormal noise model [@Shekhar2020a], the lognormal weighted evidence and visibility model [@shekhar_how_2024], the independent truncated Gaussian model [@rausch_measures_2023] based on the model specification used for the original meta-d$^\prime$/d$^\prime$ method [@Maniscalco2012; @Maniscalco2014], and the independent truncated Gaussian model based on the model specification of Hmetad [@Fleming2017a]. In addition, the `statConfR` package provides functions for estimating meta-d$^\prime$/d$^\prime$, the most widely-used measure of metacognitive efficiency, allowing both @Maniscalco2012's and @Fleming2017a's model specification. 
 Finally, the `statConfR` package includes an example data set previously published in @hellmann_simultaneous_2023, with which the functions can be  tested. 
+The `statConfR` reference manual provides documentation of each function of the latest release (https://cran.r-project.org/web/packages/statConfR/statConfR.pdf). 
 
 # Statement of need
 
@@ -60,109 +61,16 @@ and install from GitHub:
 
     devtools::install_github("ManuelRausch/StatConfR")
 
-## Usage
-
-### Data structure
-
-The package includes a demo data set from a masked orientation
-discrimination task with confidence judgments (Hellmann et al., 2023,
-Exp. 1.
-
-``` r
-library(statConfR)
-data("MaskOri")
-head(MaskOri)
-```
-
-    ##   participant stimulus correct rating diffCond trialNo
-    ## 1           1        0       1      0      8.3       1
-    ## 2           1       90       0      4    133.3       2
-    ## 3           1        0       1      0     33.3       3
-    ## 4           1       90       0      0     16.7       4
-    ## 5           1        0       1      3    133.3       5
-    ## 6           1        0       1      0     16.7       6
-
-Data should be in the form of a data.frame object columns for following
-variables:
-
-- stimulus (factor with 2 levels): The property of the stimulus which
-  defines which response is correct
-- diffCond (factor): The experimental manipulation that is expected to
-  affect discrimination sensitivity
-- correct (0-1): Indicating whether the choice was correct (1) or
-  incorrect(0).
-- rating (factor): A discrete variable encoding the decision confidence
-  (high: very confident; low: less confident)
-- participant (integer): giving the subject ID.
-
-### Fitting
-
-It is strongly recommended that if metacognitive efficiency is to be
-measured using the meta-d′/d′ method that researchers fist determine
-whether the Independent Truncated Gaussian Model, the confidence model
-implied by the meta-d′/d′ method, is an adequate description of the
-data. Using the function fitConfModel, we can fit several confidence
-models to the data of each participant. The argument
-`.parallel=TRUE`allows for parallelization over all but one available
-core.
-
[…]
