+++ README.md
@@ -8,20 +8,13 @@ assumptions of meta-d′/d′ assuming the corresponding model provides a
 better fit to the data. The following models are included:
 
 - signal detection rating model,
-
 - Gaussian noise model,
-
 - weighted evidence and visibility model,
-
 - post-decisional accumulation model,
-
 - independent Gaussian model,
-
 - independent truncated Gaussian model (the model underlying the
-  meta-d′/d′ method, see Rausch et al., 2023),
-
+- meta-d′/d′ method, see Rausch et al., 2023),
 - lognormal noise model, and
-
 - lognormal weighted evidence and visibility model.
 
 ## Mathematical description of implemented models of confidence
@@ -45,12 +38,15 @@ generate confidence, it is assumed that the confidence variable $`y`$ is
 compared to another set of criteria $`\theta_{R,i}, i=1,2,...,L-1`$,
 depending on the discrimination response $`R`$ to produce a $`L`$-step
 discrete confidence response. The different models vary in how $`y`$ is
-generated (see below). The parameters shared between all models are: -
-sensitivity parameters $`d_1, ..., d_K`$ ($`K`$: number of difficulty
-levels), - decision criterion $`c`$, - confidence criterion
-$`\theta_{-1,1}, ..., \theta_{-1,L-1},
-\theta_{1,1},  ,...,\theta_{1,L-1}`$ ($`L`$: number of confidence
-categories available for confidence ratings).
+generated (see below). The following parameters are shared between all
+models:
+
+- sensitivity parameters $`d_1, ..., d_K`$ ($`K`$: number of difficulty
+  levels),
+- decision criterion $`c`$,
+- confidence criterion $`\theta_{-1,1}, ..., \theta_{-1,L-1},
+  \theta_{1,1},  ,...,\theta_{1,L-1}`$ ($`L`$: number of confidence
+  categories available for confidence ratings).
 
 ### Signal detection rating model (SDT)
 
@@ -151,21 +147,21 @@ as free parameters.
 
 ### Logistic weighted evidence and visibility model (logWEV)
 
-The logWEV model is a combination of logN and WEV, proposed by Shekhar
+The logWEV model is a combination of logN and WEV proposed by Shekhar
 and Rahnev (2023). Conceptually, logWEV assumes that the observer
 combines evidence about decision-relevant features of the stimulus with
 the strength of evidence about choice-irrelevant features (Rausch et
 al., 2018). The model also assumes that noise affecting the confidence
 decision variable is lognormal in accordance with Shekhar and Rahnev
 (2021). According to logWEV, the confidence decision variable is $`y`$
-is equal to R × y<sup>*</sup>. y<sup>*</sup> is sampled from a lognormal
+is equal to R × y’. The variable y’ is sampled from a lognormal
 distribution with a location parameter of
 $`(1-w)\times x\times R + w \times d_k`$ and a scale parameter of
 $`\sigma`$. The parameter $`\sigma`$ quantifies the amount of
 unsystematic variability contributing to confidence judgments but not to
 the discrimination judgments. The parameter $`w`$ represents the weight
 that is put on the choice-irrelevant features in the confidence
-judgment. $`w`$ and $`\sigma`$ are free parameters.
+judgment. The parameters $`w`$ and $`\sigma`$ are free parameters.
 
 ## Measures of metacognition
 
@@ -189,10 +185,9 @@ model of confidence underlying the meta-d’/d’ method is identical to
 different versions of the independent truncated Gaussian model (Rausch
 et al., 2023), depending on whether the original model specification by
 Maniscalco and Lau (2012) or alternatively the specification by Fleming
-(2017). We strongly recommend that if metacognitive efficiency is to
-bemeasured using the meta-d′/d′ method that researchers fist determine
[…]

[…]
