+++ README.md
@@ -7,15 +7,16 @@ metacognition parameter that may serve as an alternative when the
 assumptions of meta-d′/d′ assuming the corresponding model provides a
 better fit to the data. The following models are included:
 
-- signal detection rating model,
-- Gaussian noise model,
-- weighted evidence and visibility model,
-- post-decisional accumulation model,
-- independent Gaussian model,
+- signal detection rating model (Green & Swets, 1966),
+- Gaussian noise model (Maniscalco & Lau, 2016),
+- weighted evidence and visibility model (Rausch et al., 2018),
+- post-decisional accumulation model (Rausch et al., 2018),
+- independent Gaussian model (Rausch & Zehetleitner, 2017),
 - independent truncated Gaussian model (the model underlying the
-- meta-d′/d′ method, see Rausch et al., 2023),
-- lognormal noise model, and
-- lognormal weighted evidence and visibility model.
+  meta-d′/d′ method, see Rausch et al., 2023),
+- lognormal noise model (Shekhar & Rahnev, 2021), and
+- lognormal weighted evidence and visibility model (Shekhar & Rahnev,
+  2023).
 
 ## Mathematical description of implemented models of confidence
 
@@ -54,22 +55,21 @@ According to SDT, the same sample of sensory evidence is used to
 generate response and confidence, i.e., $`y=x`$. The confidence criteria
 associated with $`R=-1`$ are more negative than the decision criterion
 $`c`$, whereas the confidence criteria associated with $`R=1`$ are more
-positive than $`c`$ (Green & Swets, 1966).
+positive than $`c`$.
 
 ### Gaussian noise model (GN)
 
 According to GN, $`y`$ is subject to additive noise and assumed to be
 normally distributed around the decision evidence value $`x`$ with a
-standard deviation $`\sigma`$, which is an additional free parameter
-(Maniscalco & Lau, 2016).
+standard deviation $`\sigma`$, which is an additional free parameter.
 
 ### Weighted evidence and visibility model (WEV)
 
 WEV assumes that the observer combines evidence about decision-relevant
 features of the stimulus with the strength of evidence about
-choice-irrelevant features to generate confidence (Rausch et al., 2018).
-Thus, the WEV model assumes that $`y`$ is normally distributed with a
-mean of $`(1-w)\times x+w \times d_k\times R`$ and standard deviation
+choice-irrelevant features to generate confidence. Thus, the WEV model
+assumes that $`y`$ is normally distributed with a mean of
+$`(1-w)\times x+w \times d_k\times R`$ and standard deviation
 $`\sigma`$. The standard deviation quantifies the amount of unsystematic
 variability contributing to confidence judgments but not to the
 discrimination judgments. The parameter $`w`$ represents the weight that
@@ -80,57 +80,56 @@ set of shared parameters.
 ### Post-decisional accumulation model (PDA)
 
 PDA represents the idea of on-going information accumulation after the
-discrimination choice (Rausch et al., 2018). The parameter $`a`$
-indicates the amount of additional accumulation. The confidence variable
-is normally distributed with mean $`x+S\times d_k\times a`$ and variance
-$`a`$. The parameter $`a`$ is fitted in addition to the shared
-parameters.
+discrimination choice. The parameter $`a`$ indicates the amount of
+additional accumulation. The confidence variable is normally distributed
+with mean $`x+S\times d_k\times a`$ and variance $`a`$. The parameter
+$`a`$ is fitted in addition to the shared parameters.
 
 ### Independent Gaussian model (IG)
 
-According to IG, $`y`$ is sampled independently from $`x`$ (Rausch &
[…]
