+++ paper.md
@@ -19,10 +19,12 @@ authors:
   orcid: "0000-0002-5805-5544"
   equal-contrib: true
   affiliation: 1, 2
+- name: Sascha Meyen
+  affiliation: 3
 - name: Sebastian Hellmann
   orcid: "0000-0001-6006-5103"
   equal-contrib: true
-  affiliation: 2, 3
+  affiliation: 2, 4
 bibliography: paper.bib
 affiliations:
 - name: "Hochschule Rhein-Waal, Fakultät Gesellschaft und Ökonomie, Germany"
@@ -30,42 +32,78 @@ affiliations:
 - name: "Katholische Universität Eichstätt-Ingolstadt, Philosophisch-pädagogische
     Fakultät, Germany"
   index: 2
-- name: TUM School of Management, Technische Universität München, Germany
+- name: Eberhard Karls Universität Tübingen, Department of Computer Science, Tübingen, Germany
   index: 3
+- name: TUM School of Management, Technische Universität München, Germany
+  index: 4
 ---
   
 # Summary
   
-We present the `statConfR` package for R, which allows researchers to conveniently fit and compare nine different static models of decision confidence applicable to binary discrimination tasks with confidence ratings: the signal detection rating model [@Green1966], the Gaussian noise model [@Maniscalco2016], the independent Gaussian model [@Rausch2017], the weighted evidence and visibility model [@Rausch2018], the lognormal noise model [@Shekhar2020a], the lognormal weighted evidence and visibility model [@shekhar_how_2024], the independent truncated Gaussian model [@rausch_measures_2023] based on the model specification used for the original meta-d$^\prime$/d$^\prime$ method [@Maniscalco2012; @Maniscalco2014], and the independent truncated Gaussian model based on the model specification of Hmetad [@Fleming2017a]. In addition, the `statConfR` package provides functions for estimating meta-d$^\prime$/d$^\prime$, the most widely-used measure of metacognitive efficiency, allowing both @Maniscalco2012's and @Fleming2017a's model specification. 
+We present the `statConfR` package for R, which allows researchers to conveniently 
+fit and compare nine different static models of decision confidence applicable
+to binary discrimination tasks with confidence ratings: 
+- the signal detection rating model [@Green1966], 
+- the Gaussian noise model [@Maniscalco2016], 
+- the independent Gaussian model [@Rausch2017], 
+- the weighted evidence and visibility model [@Rausch2018], 
+- the lognormal noise model [@Shekhar2020a], -
+- the lognormal weighted evidence and visibility model [@shekhar_how_2024], 
+- the independent truncated Gaussian model [@rausch_measures_2023] based on the model specification 
+used for the original meta-d$^\prime$/d$^\prime$ method [@Maniscalco2012; @Maniscalco2014], and 
+- the independent truncated Gaussian model based on the model specification of Hmetad [@Fleming2017a]. 
+In addition, the `statConfR` package provides functions for estimating meta-d$^\prime$/d$^\prime$, the most widely-used measure of metacognitive efficiency, allowing both @Maniscalco2012's and @Fleming2017a's model specification. 
 Finally, the `statConfR` package includes an example data set previously published in @hellmann_simultaneous_2023, with which the functions can be  tested. 
 The `statConfR` reference manual provides documentation of each function of the latest release (https://cran.r-project.org/web/packages/statConfR/statConfR.pdf). 
 
 # Statement of need
 
[…]
