### 2024-10-12 9dcbdd5
Adding Information-theoretic stuff everywhere
