+++ paper.bib
@@ -214,6 +214,23 @@
 	file = {PDF:C\:\\Users\\PPA714\\Zotero\\storage\\3HKZC26B\\Fleming (2017) HMeta-d.pdf:application/pdf},
 }
 
+@article{dayan_metacognitive_2023,
+	title = {Metacognitive {Information} {Theory}},
+	volume = {7},
+	issn = {2470-2986},
+	url = {https://direct.mit.edu/opmi/article/doi/10.1162/opmi_a_00091/116663/Metacognitive-Information-Theory},
+	doi = {10.1162/opmi_a_00091},
+	abstract = {The capacity that subjects have to rate confidence in their choices is a form of metacognition, and can be assessed according to bias, sensitivity and efficiency. Rich networks of domain-specific and domain-general regions of the brain are involved in the rating, and are associated with its quality and its use for regulating the processes of thinking and acting. Sensitivity and efficiency are often measured by quantities called meta–d0 and the M-ratio that are based on reverse engineering the potential accuracy of the original, primary, choice that is implied by the quality of the confidence judgements. Here, we advocate a straightforward measure of sensitivity, called meta–I , which assesses the mutual information between the accuracy of the subject’s choices and the confidence reports, and two normalized versions of this measure that quantify efficiency in different regimes. Unlike most other measures, meta–I -based quantities increase with the number of correctly assessed bins with which confidence is reported. We illustrate meta–I on data from a perceptual decision-making task, and via a simple form of simulated second-order metacognitive observer.},
+	language = {en},
+	urldate = {2023-09-04},
+	journal = {Open Mind},
+	author = {Dayan, Peter},
+	month = jul,
+	year = {2023},
+	pages = {392--411},
+	file = {Dayan - 2023 - Metacognitive Information Theory.pdf:C\:\\Users\\PPA714\\Zotero\\storage\\VIWIVHXC\\Dayan - 2023 - Metacognitive Information Theory.pdf:application/pdf},
+}
+
 @article{shekhar_how_2024,
 	title = {How {Do} {Humans} {Give} {Conﬁdence}? {A} {Comprehensive} {Comparison} of {Process} {Models} of {Perceptual} {Metacognition}},
 	volume = {153},
