### 2024-10-13 647bfc4
Adding a huge pile of Saschas functions
