+++ R/estimate_information.R
@@ -1,12 +1,12 @@
 #' Estimate Information
-#' 
+#'
 #' Estimate the information that a classifier's responses transmit about the
 #' true label. This uses a simple plug-in estimate where the conditional
 #' accuracy for each response is estimated as c_i. Then, each c_i determines
 #' the predictionwise (responsewise) transmitted information. The weighted
 #' sum of these predictionwise transmitted information values it the overall
 #' transmitted information.
-#' 
+#'
 #' Data can be input in three ways:
 #' - A data frame with variables "y" for true labels and "r" for
 #'   confidence-binned responses. "y" needs to contain values -1 and +1 while
@@ -19,17 +19,20 @@
 #'   labels (stimulus categories) and columns correspond to responses.
 #' - A contingency matrix with joint relative frequencies (as before but
 #'   normalized to sum up to 1).
-#' 
+#'
 #' @param x Data
-#' 
-#' @return Estimated accuracy
-#' 
+#'
+#'
 #' @return Transmitted information
+
+#' @export
 estimate_information <- function(x)
 {
   UseMethod("estimate_information")
 }
 
+
+#' @exportS3method
 estimate_information.data.frame <- function(msd)
 {
   estimated_classifier <- estimate_classifier(msd)
@@ -37,13 +40,17 @@ estimate_information.data.frame <- function(msd)
   info <- get_information(estimated_classifier)
   info
 }
+
+#' @exportS3method
 estimate_information.matrix <- function(tab)
 {
   estimated_classifier <- tab/sum(tab)
-  
+
   info <- get_information(estimated_classifier)
   info
 }
+
+#' @exportS3method
 estimate_information.table <- function(tab)
 {
   estimated_classifier <- tab/sum(tab)
