### 2024-10-13 489d4be
Adapting Saschas function for package
