+++ R/estimate_meta_I.R
@@ -33,7 +33,7 @@ estimate_meta_I <- function(x)
   UseMethod("estimate_meta_I")
 }
 
-#' @keywords internal
+#'@export
 estimate_meta_I.matrix <- function(estimated_classifier)
 {
   estimated_classifier <- estimated_classifier/sum(estimated_classifier)
@@ -48,7 +48,7 @@ estimate_meta_I.matrix <- function(estimated_classifier)
   meta_I
 }
 
-#' @keywords internal
+#' @export
 estimate_meta_I.data.frame <- function(msd)
 {
   estimated_classifier <- estimate_classifier(msd)
@@ -57,7 +57,7 @@ estimate_meta_I.data.frame <- function(msd)
   meta_I
 }
 
-#' @keywords internal
+#' @export
 estimate_meta_I.table <- function(counts_table)
 {
   estimated_classifier <- counts_table/sum(counts_table)
