### 2024-10-14 ca2f0e1
Completing documnetation with metainfo measures
