+++ README.md
@@ -1,11 +1,15 @@
 The `statConfR` package provides functions to fit static models of
 decision-making and confidence derived from signal detection theory for
-binary discrimination tasks, as well as meta-d′/d′ as measures of
-metacognition. Fitting models of confidence can be used to test the
-assumptions underlying meta-d′/d′. Several models include a
-metacognition parameter that may serve as an alternative when the
-assumptions of meta-d′/d′ assuming the corresponding model provides a
-better fit to the data. The following models are included:
+binary discrimination tasks, meta-d′/d′, a wide-spread measure of
+metacognitive efficiency, as well as information-theoretic measures of
+metacognitive sensitivity and efficiency.
+
+Fitting models of confidence can be used to test the assumptions
+underlying meta-d′/d′. Several static models of decision-making and
+confidence include a metacognition parameter that may serve as an
+alternative when the assumptions of meta-d′/d′ assuming the
+corresponding model provides a better fit to the data. The following
+models are included:
 
 - signal detection rating model (Green & Swets, 1966),
 - Gaussian noise model (Maniscalco & Lau, 2016),
@@ -59,23 +63,26 @@ positive than $`c`$.
 
 ### Gaussian noise model (GN)
 
-According to GN, $`y`$ is subject to additive noise and assumed to be
-normally distributed around the decision evidence value $`x`$ with a
-standard deviation $`\sigma`$, which is an additional free parameter.
+Conceptually, the Gaussian noise model reflects the idea that confidence
+is informed by the same sensory evidence as the task decision, but
+confidence is affected by additive Gaussian noise. According to GN,
+$`y`$ is subject to additive noise and assumed to be normally
+distributed around the decision evidence value $`x`$ with a standard
+deviation $`\sigma`$, which is an additional free parameter.
 
 ### Weighted evidence and visibility model (WEV)
 
-WEV assumes that the observer combines evidence about decision-relevant
-features of the stimulus with the strength of evidence about
-choice-irrelevant features to generate confidence. Thus, the WEV model
-assumes that $`y`$ is normally distributed with a mean of
-$`(1-w)\times x+w \times d_k\times R`$ and standard deviation
-$`\sigma`$. The standard deviation quantifies the amount of unsystematic
-variability contributing to confidence judgments but not to the
-discrimination judgments. The parameter $`w`$ represents the weight that
-is put on the choice-irrelevant features in the confidence judgment. The
-parameters $`w`$ and $`\sigma`$ are free parameters in addition to the
-set of shared parameters.
+Conceptually, the WEV model reflects the idea that the observer combines
+evidence about decision-relevant features of the stimulus with the
+strength of evidence about choice-irrelevant features to generate
+confidence. For this purpose, WEV assumes that $`y`$ is normally
+distributed with a mean of $`(1-w)\times x+w \times d_k\times R`$ and
+standard deviation $`\sigma`$. The standard deviation quantifies the
+amount of unsystematic variability contributing to confidence judgments
+but not to the discrimination judgments. The parameter $`w`$ represents
+the weight that is put on the choice-irrelevant features in the
+confidence judgment. The parameters $`w`$ and $`\sigma`$ are free
+parameters in addition to the set of shared parameters.
[…]
