+++ R/get_accuracy.R
@@ -1,4 +1,4 @@
-rget_accuracy <- function(classifier)
+get_accuracy <- function(classifier)
 {
   accuracy <- sum( apply(classifier, 2, max) )
   accuracy
