### 2024-10-14 cc6bf44
S3 consistency
make sure that S3methods have all the same argument
