+++ R/estimate_accuracy.R
@@ -1,31 +1,30 @@
-# This function is not getting documented because I think that the use case to ignore
-# the mapping intended by the classifier so rare that the chance is much higher that
-# someone will mess up using this function! The function is available internally, though.
-estimate_accuracy <- function(x)
-{
+# R cmd check throws a warning if S3generic and S3methods do not accept identical arguments.
+# This may cause problems when submitting the package to CRAN later.
+
+estimate_accuracy <- function(x){
   UseMethod("estimate_accuracy")
 }
 
 #' @export
-estimate_accuracy.matrix <- function(estimated_classifier)
+estimate_accuracy.matrix <- function(x)
 {
-  estimated_classifier <- estimated_classifier/sum(estimated_classifier)
+  estimated_classifier <- x/sum(x)
 
   acc <- get_accuracy(estimated_classifier)
   acc
 }
 #'@export
-estimate_accuracy.data.frame <- function(msd)
+estimate_accuracy.data.frame <- function(x)
 {
-  estimated_classifier <- estimate_classifier(msd)
+  estimated_classifier <- estimate_classifier(x)
 
   acc <- estimate_accuracy.matrix(estimated_classifier)
   acc
 }
 #'@export
-estimate_accuracy.table <- function(counts_table)
+estimate_accuracy.table <- function(x)
 {
-  estimated_classifier <- counts_table/sum(counts_table)
+  estimated_classifier <- x/sum(x)
 
   acc <- estimate_accuracy.matrix(estimated_classifier)
   acc

[…]
