+++ paper.md
@@ -52,10 +52,12 @@ binary discrimination tasks with confidence ratings:
 - the independent truncated Gaussian model [@rausch_measures_2023] based on the model specification 
 used for the original meta-d$^\prime$/d$^\prime$ method [@Maniscalco2012; @Maniscalco2014], and 
 - the independent truncated Gaussian model based on the model specification of Hmetad [@Fleming2017a]. 
-In addition, the `statConfR` package provides functions for estimating two different 
+In addition, the `statConfR` package provides functions for estimating different 
 kinds of measures of metacognition: 
 - meta-d$^\prime$/d$^\prime$, the most widely-used measure of metacognitive efficiency, allowing both @Maniscalco2012's and @Fleming2017a's model specification, 
-- Information-theoretic measures of metacognitive sensitivity and metacognitive efficiency [@dayan_metacognitive_2023]. 
+- Information-theoretic measures [@dayan_metacognitive_2023], including
+  - meta-I, an information-theoretic measures of metacognitive sensitivity, 
+  - $meta-I_{1}^{r}$ and $meta-I_{2}^{r}$, two measures of metacognitive efficiency. 
 Finally, the `statConfR` package includes an example data set previously published in @hellmann_simultaneous_2023, with which the functions can be  tested. 
 The `statConfR` reference manual provides documentation of each function of the latest release (https://cran.r-project.org/web/packages/statConfR/statConfR.pdf). 
 
@@ -92,7 +94,7 @@ The `statConfR` package also provides a faithful implementation of meta-d$^\prim
 which has been originally implemented in MATLAB [@Maniscalco2012]. Fleming provides MATLAB and R code for Hmetad, 
 a Bayesian hierarchical version of meta-d$^\prime$/d$^\prime$ [@Fleming2017a], 
 but notably he specifies the model slightly differently as in the original meta-d$^\prime$/d$^\prime$ [@rausch_measures_2023]. 
-To our knowledge, information-theoretic
+To our knowledge, there has been no open software available to estimate information-theoretic measures of metacognition up to now. 
 
 An important limitation of the models implemented in `statConfR` is that the dynamics of the decision process are not taken into account.
 This is a problem because confidence judgments are related to the dynamics of decision making [@hellmann_confidence_2024; @Pleskac2010; @Rahnev2020].
