+++ README.rmd
@@ -20,7 +20,9 @@ gitbranch <- "main/"
 
 The `statConfR` package provides functions to fit static models of
 decision-making and confidence derived from signal detection theory for
-binary discrimination tasks, meta-d′/d′, a wide-spread measure of metacognitive efficiency, as well as information-theoretic measures of metacognitive sensitivity and efficiency. 
+binary discrimination tasks, meta-d′/d′, a wide-spread measure of metacognitive efficiency, 
+meta-I, an information-theoretic measures of metacognitive sensitivity, 
+as well as $meta-I_{1}^{r}$ and $meta-I_{2}^{r}$, two information-theoretic measures of metacognitive efficiency. 
 
 Fitting models of confidence can be used to test the assumptions underlying
 meta-d′/d′. Several static models of decision-making and confidence include a metacognition parameter that may
@@ -162,14 +164,15 @@ because both are measured on the same scale. Meta-d′ can be compared against t
 ### Information-theoretic measures of metacognition
 
 Dayan (2023) proposed several measures of metacognition based on quantities of information theory. 
-* Meta-I is a measure of metacognitive sensitivity defined as the mutual information between the confidence and accuracy and is calculated as the transmitted information minus the minimal information given the accuracy,
-$$meta-I = I(Y; \hat{Y}, C) - I(Y; \hat{Y})$$. 
+
+- Meta-I is a measure of metacognitive sensitivity defined as the mutual information between the confidence and accuracy and is calculated as the transmitted information minus the minimal information given the accuracy,
+$$meta-I = I(Y; \hat{Y}, C) - I(Y; \hat{Y}).$$ 
 This is equivalent to Dayan's formulation where meta-I is the information that confidences transmit about the correctness of a response: 
 $$meta-I = I(Y = \hat{Y}; C)$$
-* Meta-$I_{1}^{r}$ is meta-I  normalized by the value of meta-I expected assuming 
+- Meta-$I_{1}^{r}$ is meta-I  normalized by the value of meta-I expected assuming 
 a signal detection model (Green & Swets, 1966) with Gaussian noise, based on calculating the sensitivity index d':
 $$meta-I_{1}^{r} = meta-I / meta-I(d')$$
- * Meta-$I_{2}^{r}$ is meta-I normalized by its theoretical upper bound, which is the information entropy of accuracy,  $H(Y = \hat{Y})$:
+- Meta-$I_{2}^{r}$ is meta-I normalized by its theoretical upper bound, which is the information entropy of accuracy,  $H(Y = \hat{Y})$:
 $$meta-I_{2}^{r} = meta-I / H(Y = \hat{Y})$$
 
 Notably, Dayan (2023) pointed out that a liberal or conservative use of the confidence levels will affected the mutual information and thus all information-theoretic measures of metacognition. 
@@ -209,7 +212,8 @@ head(MaskOri)
 The function `fitConfModels` allows the user to fit several confidence
 models separately to the data of each participant. 
 The data should be provided via the argument `.data` in the form of a data.frame 
-object with the following variables in separate columns: 
+object with the following variables in separate columns:
+
 -   stimulus (factor with 2 levels): The property of the stimulus which
     defines which response is correct
 -   diffCond (factor): The experimental manipulation that is expected to
@@ -240,7 +244,7 @@ It can be seen that the independent truncated Gaussian model is consistently out
 
 ### Visualization 
 
[…]
