### 2024-10-21 4efb91e
Adapted Information-Theoretic Measures of Metacognition to Original Convention
Summarized the meta-I measures in one function, estimateMetaI(), that  takes the same data frame format as the original implementations, for example fitConf(). Also added bias reduction estimators. Streamlined and cleaned the corresponding code.
