+++ R/get_pointwise_transmitted_information.R
@@ -1,3 +1,0 @@
-get_pointwise_transmitted_information <- pti <- function(c, prior){
-  H(prior) - H(c)
-}
