### 2024-10-21 05dd3ec
Small addition about RMI (Sascha's measure) in the Readme
