### 2024-10-21 bdc397d
Added myself to author contributions (just to fill the gap, not to overstate what I did).
