### 2024-10-22 7dd41a1
Small update in the ?estimateMetaI
Previously, the order of measures in the documentation was different from that in the output data.frame. Now it is sorted and makes clear that the recent addition to the meta-I-zoo, RMI, is the last of these measures and not discussed by Dayan.
