### 2024-10-22 c85a34a
Updated documentation of estimateMetaI
