+++ paper.md
@@ -20,6 +20,7 @@ authors:
   equal-contrib: true
   affiliation: 1, 2
 - name: Sascha Meyen
+  orcid: "0000-0001-6928-4126"
   affiliation: 3
 - name: Sebastian Hellmann
   orcid: "0000-0001-6006-5103"
