+++ R/estimateMetaI.R
@@ -1,41 +1,41 @@
-#' @title Estimate Metacognitive Information Theory Measures
+#' @title Estimate Measures of Metacognition from Information Theory
 
 #' @description `estimateMetaI` estimates meta-\eqn{I}, an information-theoretic
 #'  measure of metacognitive sensitivity proposed by Dayan (2023), as well as
-#'  derived measures like it including meta-\eqn{I_{1}^{r}} and Meta-\eqn{I_{2}^{r}}.
+#'  similar derived measures, including meta-\eqn{I_{1}^{r}} and Meta-\eqn{I_{2}^{r}}.
 #'  These are different normalizations of meta-\eqn{I}:
-#' - meta-\eqn{I_{1}^{r}} normalizes by the meta-\eqn{I} that would be
+#' - Meta-\eqn{I_{1}^{r}} normalizes by the meta-\eqn{I} that would be
 #'   expected from an underlying normal distribution with the same
 #'   sensitivity.
-#' - meta-\eqn{I_{1}^{r\prime}} is a variant not discussed by Dayan
+#' - Meta-\eqn{I_{1}^{r\prime}} is a variant of meta-\eqn{I_{1}^{r}} not discussed by Dayan
 #'   (2023) which normalizes by the meta-\eqn{I} that would be expected from
-#'   an underlying normal distribution with the same accuracy.(This is
+#'   an underlying normal distribution with the same accuracy (this is
 #'   similar to the sensitivity approach but without considering variable
-#'   thresholds.)
-#' - meta-\eqn{I_{2}^{r}} normalizes by the maximum amount of meta-\eqn{I}
+#'   thresholds).
+#' - Meta-\eqn{I_{2}^{r}} normalizes by the maximum amount of meta-\eqn{I}
 #'   which would be reached if all uncertainty about the stimulus was removed.
-#'  Additionally, one more measure that was not discussed by Dayan (2023) is
-#'  provided.
 #' - \eqn{RMI} normalizes meta-\eqn{I} by the range of its possible
-#'    values and therefore scales between 0 and 1.
-#'  All measures are also calculated with a bias-reduced variant for which the
+#'    values and therefore scales between 0 and 1. RMI is a novel measure not discussed by Dayan (2023).
+#'
+#'  All measures can be calculated with a bias-reduced variant for which the
 #'  observed frequencies are taken as underlying probability distribution to
-#'  estimate the sampling bias. The estimated bias is then subtraced from the
+#'  estimate the sampling bias. The estimated bias is then subtracted from the
 #'  initial measures. This approach uses Monte-Carlo simulations and is
 #'  therefore not deterministic (values can vary from one evaluation of the
-#'  function to the next). However, it is a simple way to reduce the bias
+#'  function to the next). However, this is a simple way to reduce the bias
 #'  inherent in these measures.
 
 #' @details
-#' Meta-\deqn{I} is defined as the mutual information between the confidence and
+#' Meta-\eqn{I} is defined as the mutual information between the confidence and
 #' accuracy and is calculated as the transmitted information minus the
 #' minimal information given the accuracy,
-#' meta-\deqn{I = I(Y; \hat{Y}, C) - I(Y; \hat{Y})}.
+#' \deqn{meta-I = I(Y; \hat{Y}, C) - I(Y; \hat{Y}).}
 #' This is equivalent to Dayan's formulation where meta-I is the information
-#' that confidences transmit about the correctness of a response,
-#' meta-\deqn{I = I(Y = \hat{Y}; C).}
-#' The other measures here are different normalizations of this Meta-\deqn
-#' {I}. It should be noted that Dayan (2023) pointed out that a liberal or
[…]
