### 2024-10-23 bd42f97
Adding a manual
