### 2024-10-25 ce5d2d4
Updated handling of extreme hit and false alarm rates. Also unified debiasing simulation so that not each measure is simulated separately (for speedup).
