### 2024-10-31 5b7cd0e
create plotting function
