### 2024-11-29 b5fe057
Adding code for figures
