+++ R/plotConfModelFit.R
@@ -97,8 +97,8 @@ plotConfModelFit <- function(data, fitted_pars, model = NULL){
 
   AggDist <-
     plyr::ddply(data,
-          ~  diffCond * rating * stimulus * correct * participant, #,
-          plyr::summarise, p = length(rating),  .drop=FALSE)
+                ~  diffCond * rating * stimulus * correct * participant, #,
+                plyr::summarise, p = length(rating),  .drop=FALSE)
 
   AggDist <- plyr::ddply(AggDist, ~ diffCond * stimulus,
                          transform, N = sum(p))
@@ -111,7 +111,7 @@ plotConfModelFit <- function(data, fitted_pars, model = NULL){
   levels(AggDist$correct)  <-
     c("incorrect", "correct")
 
-   # 2. aggregate across subjects
+  # 2. aggregate across subjects
   AggDist <-
     Rmisc::summarySEwithin(AggDist , measurevar = "p",
                            withinvars = c("diffCond", "correct", "rating", "stimulus"),
@@ -119,9 +119,36 @@ plotConfModelFit <- function(data, fitted_pars, model = NULL){
                            na.rm = TRUE, .drop = TRUE)
   AggDist$rating <- as.numeric(AggDist$rating)
   levels(AggDist$stimulus) <- c("S = -1", "S = 1")
+  AggDist$diffCond <- factor(as.numeric(AggDist$diffCond)) # diffCond should code the order of difficulty levels
   levels(AggDist$diffCond) <- paste("K =", as.numeric(levels(AggDist$diffCond)))
 
-  # 3) create a plot with the observed data
+  # 4) create the prediction from model fit
+
+  predictDataFun <- switch(model,
+                           'WEV' = predictDataWEV,
+                           'logWEV' = predictDataLogWEV,
+                           'SDT' = predictDataSDT,
+                           'IG' = predictData2Chan,
+                           'ITGc' = predictDataIndTruncF,
+                           'ITGcm' = predictDataIndTruncML,
+                           'GN' = predictDataNoisy,
+                           'PDA' = predictDataISDT,
+                           'logN' = predictDataLognorm)
+  myParams <- fitted_pars[fitted_pars$model == model,]
+  predictedData <- plyr::ddply(myParams, ~participant, predictDataFun)
+  predictedData$correct <-
+    factor(ifelse(predictedData$stimulus==predictedData$response, "correct", "incorrect"))
+
+  AggPredictedData <-
+    aggregate(p ~ stimulus*diffCond*rating*correct, predictedData, mean)
+  AggPredictedData$stimulus <- factor(AggPredictedData$stimulus)
+  levels(AggPredictedData$stimulus) <- c("S = -1", "S = 1")
+  AggPredictedData$diffCond <- factor(AggPredictedData$diffCond)
+  levels(AggPredictedData$diffCond) <-
[…]
