### 2024-11-30 0513393
Adding plotting capability
