+++ NAMESPACE
@@ -4,10 +4,14 @@ export(estimateMetaI)
 export(fitConf)
 export(fitConfModels)
 export(fitMetaDprime)
+export(plotConfModelFit)
 export(simConf)
+import(ggplot2)
 import(parallel)
+importFrom(Rmisc,summarySEwithin)
 importFrom(plyr,ddply)
 importFrom(plyr,mdply)
+importFrom(plyr,summarise)
 importFrom(stats,dnorm)
 importFrom(stats,integrate)
 importFrom(stats,optim)

[…]
