+++ paper.bib
@@ -117,6 +117,23 @@
 	file = {PDF:C\:\\Users\\PPA714\\Zotero\\storage\\S5EC6DSC\\Rausch (2017) Should metacognition be measured by logistic regression.pdf:application/pdf},
 }
 
+@article{Rigoux2014,
+	title = {Bayesian model selection for group studies - {Revisited}},
+	volume = {84},
+	issn = {10959572},
+	url = {http://dx.doi.org/10.1016/j.neuroimage.2013.08.065},
+	doi = {10.1016/j.neuroimage.2013.08.065},
+	abstract = {In this paper, we revisit the problem of Bayesian model selection (BMS) at the group level. We originally addressed this issue in Stephan et al. (2009), where models are treated as random effects that could differ between subjects, with an unknown population distribution. Here, we extend this work, by (i) introducing the Bayesian omnibus risk (BOR) as a measure of the statistical risk incurred when performing group BMS, (ii) highlighting the difference between random effects BMS and classical random effects analyses of parameter estimates, and (iii) addressing the problem of between group or condition model comparisons. We address the first issue by quantifying the chance likelihood of apparent differences in model frequencies. This leads to the notion of protected exceedance probabilities. The second issue arises when people want to ask "whether a model parameter is zero or not" at the group level. Here, we provide guidance as to whether to use a classical second-level analysis of parameter estimates, or random effects BMS. The third issue rests on the evidence for a difference in model labels or frequencies across groups or conditions. Overall, we hope that the material presented in this paper finesses the problems of group-level BMS in the analysis of neuroimaging and behavioural data. © 2013 Elsevier Inc.},
+	journal = {NeuroImage},
+	author = {Rigoux, L. and Stephan, K. E. and Friston, K. J. and Daunizeau, J.},
+	year = {2014},
+	pmid = {24018303},
+	note = {Publisher: Elsevier Inc.},
+	keywords = {Between-condition comparison, Between-group comparison, DCM, Exceedance probability, Mixed effects, Random effects, Statistical risk},
+	pages = {971--985},
+	file = {PDF:C\:\\Users\\PPA714\\Zotero\\storage\\YKPFAD78\\Rigoux (2014) Model selection.pdf:application/pdf},
+}
+
 @article{Rausch2018,
 	title = {Confidence in masked orientation judgments is informed by both evidence and visibility},
 	volume = {80},
@@ -350,4 +367,24 @@
 	publisher = {Routledge},
 	author = {Hautus, Michael J and Macmillan, Neil A and Creelman, C Douglas},
 	year = {2021},
+}
+
+@article{daunizeau_vba_2014,
+	title = {{VBA}: {A} {Probabilistic} {Treatment} of {Nonlinear} {Models} for {Neurobiological} and {Behavioural} {Data}},
+	volume = {10},
+	issn = {1553-7358},
+	shorttitle = {{VBA}},
+	url = {https://dx.plos.org/10.1371/journal.pcbi.1003441},
+	doi = {10.1371/journal.pcbi.1003441},
[…]
