+++ README.md
@@ -36,7 +36,7 @@ different kinds of measures of metacognition:
   - RMI, a novel measure of metacognitive accuracy, also derived from
     information theory.
 
-## 1.1 Mathematical description of implemented generative models of confidence
+# 2 Mathematical description of implemented generative models of confidence
 
 The models included in the statConfR package are all based on signal
 detection theory (Green & Swets, 1966). It is assumed that participants
@@ -67,7 +67,7 @@ models:
   \theta_{1,1},  ,...,\theta_{1,L-1}`$ ($`L`$: number of confidence
   categories available for confidence ratings).
 
-### 1.1.1 Signal detection rating model (SDT)
+## 2.1 Signal detection rating model (SDT)
 
 According to SDT, the same sample of sensory evidence is used to
 generate response and confidence, i.e., $`y=x`$. The confidence criteria
@@ -75,7 +75,7 @@ associated with $`R=-1`$ are more negative than the decision criterion
 $`c`$, whereas the confidence criteria associated with $`R=1`$ are more
 positive than $`c`$.
 
-### 1.1.2 Gaussian noise model (GN)
+## 2.2 Gaussian noise model (GN)
 
 Conceptually, the Gaussian noise model reflects the idea that confidence
 is informed by the same sensory evidence as the task decision, but
@@ -84,7 +84,7 @@ $`y`$ is subject to additive noise and assumed to be normally
 distributed around the decision evidence value $`x`$ with a standard
 deviation $`\sigma`$, which is an additional free parameter.
 
-### 1.1.3 Weighted evidence and visibility model (WEV)
+## 2.3 Weighted evidence and visibility model (WEV)
 
 Conceptually, the WEV model reflects the idea that the observer combines
 evidence about decision-relevant features of the stimulus with the
@@ -98,7 +98,7 @@ the weight that is put on the choice-irrelevant features in the
 confidence judgment. The parameters $`w`$ and $`\sigma`$ are free
 parameters in addition to the set of shared parameters.
 
-### 1.1.4 Post-decisional accumulation model (PDA)
+## 2.4 Post-decisional accumulation model (PDA)
 
 PDA represents the idea of on-going information accumulation after the
 discrimination choice. The parameter $`a`$ indicates the amount of
@@ -106,7 +106,7 @@ additional accumulation. The confidence variable is normally distributed
 with mean $`x+S\times d_k\times a`$ and variance $`a`$. The parameter
 $`a`$ is fitted in addition to the shared parameters.
 
-### 1.1.5 Independent Gaussian model (IG)
+## 2.5 Independent Gaussian model (IG)
 
 According to IG, the information used for confidence judgments is
 generated independently from the sensory evidence used for the task
@@ -117,7 +117,7 @@ parameter $`m`$ represents the amount of information available for
 confidence judgment relative to amount of evidence available for the
 discrimination decision and can be smaller as well as greater than 1.
 
-### 1.1.6 Independent truncated Gaussian model: HMetad-Version (ITGc)
+## 2.6 Independent truncated Gaussian model: HMetad-Version (ITGc)
 
 Conceptually, the two ITG models just as IG are based on the idea that
 the information used for confidence judgments is generated independently
@@ -137,7 +137,7 @@ the amount of information available for confidence judgments relative to
 amount of evidence available for discrimination decisions and can be
 smaller as well as greater than 1.
 
-### 1.1.7 Independent truncated Gaussian model: Meta-d’-Version (ITGcm)
+## 2.7 Independent truncated Gaussian model: Meta-d’-Version (ITGcm)
 
 According to the version of the ITG consistent with the original meta-d’
[…]
