+++ README.rmd
@@ -43,7 +43,7 @@ kinds of measures of metacognition:
   - meta-$I_{1}^{r\prime}$, a novel variant of $meta-I_{1}^{r}$,
   - RMI, a novel measure of metacognitive accuracy, also derived from information theory.  
   
-## Mathematical description of implemented generative models of confidence
+# Mathematical description of implemented generative models of confidence
 The models included in the statConfR package are all based on signal detection theory (Green & Swets, 1966). 
 It is assumed that participants select a binary discrimination response $R$ about a stimulus $S$.
 Both $S$ and $R$ can be either -1 or 1. $R$ is considered correct if $S=R$.
@@ -66,19 +66,19 @@ The following parameters are shared between all models:
 - confidence criterion $\theta_{-1,1}, ..., \theta_{-1,L-1}, 
 \theta_{1,1},  ,...,\theta_{1,L-1}$ ($L$: number of confidence categories available for confidence ratings).
 
-### Signal detection rating model (SDT)
+## Signal detection rating model (SDT)
 According to SDT, the same sample of sensory evidence is used to generate response and confidence, i.e.,
 $y=x$. The confidence criteria associated with $R=-1$ are more negative than the 
 decision criterion $c$, whereas the confidence criteria associated with $R=1$ are more positive than $c$.
 
-### Gaussian noise model (GN)
+## Gaussian noise model (GN)
 Conceptually, the Gaussian noise model reflects the idea that confidence is informed by the same sensory evidence
 as the task decision, but confidence is affected by additive Gaussian noise.
 According to GN, $y$ is subject to additive noise and assumed to be normally distributed
 around the decision evidence value $x$ with a standard deviation $\sigma$, 
 which is an additional free parameter.
 
-### Weighted evidence and visibility model (WEV)
+## Weighted evidence and visibility model (WEV)
 Conceptually, the WEV model reflects the idea that the observer combines evidence about decision-relevant features 
 of the stimulus with the strength of evidence about choice-irrelevant features 
 to generate confidence. For this purpose, WEV assumes that $y$ is normally 
@@ -89,14 +89,14 @@ The parameter $w$ represents the weight that is put on the choice-irrelevant
 features in the confidence judgment. The parameters $w$ and $\sigma$ are free parameters in
 addition to the set of shared parameters.
 
-### Post-decisional accumulation model (PDA)
+## Post-decisional accumulation model (PDA)
 PDA represents the idea of on-going information accumulation after the
 discrimination choice. The parameter $a$ indicates the amount of additional
 accumulation. The confidence variable is normally distributed with mean
 $x+S\times d_k\times a$ and variance $a$. The parameter $a$ is fitted in addition to the shared
 parameters.
 
-### Independent Gaussian model (IG)
+## Independent Gaussian model (IG)
 According to IG, the information used for confidence judgments is generated independently from the sensory evidence used for the task decision. 
 For this purpose, it is assumed that $y$ is sampled independently from $x$. 
 The variable $y$ is normally distributed with a mean of $a\times d_k$ and variance
@@ -104,7 +104,7 @@ of 1. The additional parameter $m$ represents the amount of information availabl
 relative to amount of evidence available for the discrimination decision and can 
 be smaller as well as greater than 1.
 
-### Independent truncated Gaussian model: HMetad-Version (ITGc)
+## Independent truncated Gaussian model: HMetad-Version (ITGc)
 Conceptually, the two ITG models just as IG are based on the idea that the information used for confidence judgments is generated independently from the sensory evidence used for the task decision. However, in contrast to IG,  the two ITG models also reflect a form of confirmation bias in so far as 
 it is not possible to collect information that contradicts the original decision. 
[…]
