### 2024-12-08 a6a1be1
verious updates to the documentation
