+++ man/plotConfModelFit.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/plotConfModelFit.R
 \name{plotConfModelFit}
 \alias{plotConfModelFit}
-\title{Plot the prediction of fitted parameters of one model of confidence over the corresponding real data}
+\title{Plot the prediction of fitted parameters of one model of confidence over the corresponding  data}
 \usage{
 plotConfModelFit(data, fitted_pars, model = NULL)
 }
@@ -26,44 +26,49 @@ determined by the different values of this column)
 }}
 
 \item{fitted_pars}{a \code{data.frame} with one row for each participant and model parameters in different columns.
-fitted_pars also may contain a column called  \code{model}specifying the model to be visualized.
+fitted_pars also may contain a column called \code{model} specifying the model to be visualized.
 If there is no model column in data or if there are multiple models in fitted_pars,
 it is necessary to specify the model argument.}
 
-\item{model}{\code{character}. See \code{\link{fitConf}} for all available models}
+\item{model}{\code{character}. See \code{\link{fitConfModels}} for all available models}
 }
 \value{
-a \code{data.frame} with one row for each combination of model and
-participant. There are different columns for the model, the participant ID, and one
-one column for each estimated model parameter (parameters
-not present in a specific model are filled with NAs)
-
-#' @examples1. Select two subjects from the masked orientation discrimination experiment
-
-data <- subset(MaskOri, participant \%in\% c(1:2))
-head(data)2. Fit some models to each subject of the masked orientation discrimination experiment
-
-\donttest{
-  # Fitting several models to several subjects takes quite some time
-  # (about 10 minutes per model fit per participant on a 2.8GHz processor
-  # with the default values of nInits and nRestart).
-  # If you want to fit more than just two subjects,
-  # we strongly recommend setting .parallel=TRUE
-  Fits <- fitConfModels(data, models = "ITGc", .parallel = FALSE)
-}3. Plot the predicted probabilies based on model and fitted parameter over the observed relative frequencies.
-
-\donttest{
-  # Fitting several models to several subjects takes quite some time
-  # (about 10 minutes per model fit per participant on a 2.8GHz processor
-  # with the default values of nInits and nRestart).
-  # If you want to fit more than just two subjects,
-  # we strongly recommend setting .parallel=TRUE
-  myPlottedFit <- plotConfModelFit(data, Fits)
-  myPlottedFit
-}
+a \code{ggplot} object with empirically observed distribution of responses and confidence ratings
+as bars on the x-axis as a function of discriminability (in the rows) and stimulus
+(in the columns). Superimposed on the empirical data,
+the plot also shows the prediction of one selected model as dots.
 }
 \description{
[…]
