### 2024-12-08 #6966 @ManuelRausch
> Documentation: I share the same feeling as [@christinamaher](https://github.com/christinamaher) that the package can benefit greatly from more documentation. The way that the paper is currently written and the README file makes me feel that the author assumes a relatively high degree of expertise in the related research domain in the audience, which may or may not be true. Therefore, I think it would be great to provide a step-by-step walkthrough example in the paper. This can be the current example use described in README, but probably with more text descriptions e.g., what the output means; what are some functions that can be used; what are the orders that the functions should be used.

According to the JOSS website, JOSS papers are expected to have metadata, a statement of need, Summary, Acknowledgements, and References sections. Software documentation should not be in the paper and instead should be outlined in the software documentation (https://joss.readthedocs.io/en/latest/paper.html). Although I saw examples papers where this rule did not seem to be strictly enforced, we feel that it is better to keep paper and documentation separate: Although I am not planning to change the user interface, sooner or later, there will be updates to the package. The software documentation can be more promptly updated than the paper.

However, we agree that a step-by-step walkthrough example in the paper is useful. For this purpose, we have worked out the example in the README file (see section 5). The usage example demonstrates the workflow in which we envision that the functions are used, and what the output of each function means.
