### 2024-12-08 #6966 @ManuelRausch
> Relatedly, I notice that on the function help page there are descriptions related to the input data format as well as detailed description of each model. It would be great to make it explicit (e.g., mention that the users can refer to XXX page, or paraphrase them concisely in the paper/README

In the new version of the readme, we describe how the documentation can be accessed (section 6).  In addition, the readme now includes a mathematical description of each model  (section 2). Finally, the revised usage example in the readme describes the inputs and outputs of all function we expected the user to interact with (sections 5.2 – 5.4).
