### 2024-12-08 #6966 @ManuelRausch
> Is it possible for the audience to implement their own models? Is there any guideline on how to do that, and is there any platform for people to share the model with each other? I understand that this maybe too much to ask for, but either way (allowing community contribution or not), it may worth spoiling some ink describing whether the users can implement their customized models, whether this it is recommended, and what are some tips that can be shared.

In the new version of the readme, there is a step-by-step instruction how to implement new models (section 6). We very much welcome contributions of new models by the community, although we recommend implementing custom-built models only for researchers with experience in cognitive modelling.
