### 2024-12-08 #6966 @ManuelRausch
> When I was running the example code, the function fitConfModels took a while to run. I am not sure whether this is normal/my computer is too old. Either way, it may worth adding some description on README for the running speed so that the users can calibrate their expectation (I was anxiously checking whether my R is still working since there was no output for a while).

We are very sorry for the inconvenience our code caused on your computer. We added an estimate how long the code may need to run to the README file (section 5.2) as well as to the documentation of fitConfModels.
