### 2024-12-08 #6966 @ManuelRausch
> General Checks - Human Research Approvals: The dataset provided is from a human experiment. Although it appears to have been previously published, the documentation does not provide details regarding informed consent for data sharing or ethical approvals.

In the new version of the paper, we state in the acknowledgment that the original study had been approved by the ethics committee of the Catholic University of Eichstätt-Ingolstadt. All participants who contributed to the data set provided written consent for participating in the experiment and for publishing their anonymized data in scientific repositories.
