### 2024-12-08 #6966 @ManuelRausch
> Documentation - Functionality Documentation: The code would benefit significantly from improved documentation.
Function Documentation: Each function is lacking explanation of its purpose, parameters, and outputs.

Each function with which the user is expected to interact has a documentation, describing the purpose of the function, its parameters, and its output. The documentation is accessible after installation of the package by typing  ?functionname into the console. In the new version of the readme, we describe how the documentation can be found (section 6).
