### 2024-12-08 #6966 @ManuelRausch
> Software Paper - Summary: The paper lacks a high-level summary of the package’s primary use-case. Including this summary would help readers understand the main objectives and applications of the package

A high-level summary of the package primary use case is found in the section “summary” of the paper.
