### 2024-12-08 #6966 @ManuelRausch
> Model Descriptions: Authors should consider providing a detailed description of the models included in the package and an explain for the rationale behind the choice of models and the fitting procedures used.

We provide a detailed descriptions of all models in the readme file (section 2) as well
as in the documentation of the function fitConfModels. The description includes the conceptual idea underlying each model. We also provide details about the fitting algorithm in both readme (section 5.2) and documentation (?fitConfModels).
A detailed description of each model is not included in the paper because the paper would become far to long for JOSS.
