### 2024-12-08 #6966 @ManuelRausch
> Visual Aids: Incorporate figures or schematics that visualize the models and the results would greatly enhance the paper's readability and package's usability. Plotting Utilities: Consider adding plotting utilities to the package to facilitate visualization of results.

We added the function plotConfModelFit to the package that plots the predicted distribution of discrimination responses and confidence ratings created from the output produced by fitConfModels and overlays the predicted distribution over the data to which the model parameters were fitted (see section 5.3 of the readme file).
