### 2025-01-06 #6966 @samhforbes
Hi @christinamaher and @haoxue-fan happy new year! Just wanted to check whether your queries had been answered by this latest update.
