+++ paper.md
@@ -23,7 +23,7 @@ authors:
   orcid: "0000-0001-6928-4126"
   affiliation: 3
 - name: Sebastian Hellmann
-  orcid: "0000-0001-6006-5103"
+  orcid: "0000-0002-3621-6343"
   equal-contrib: true
   affiliation: 2, 4
 bibliography: paper.bib
