### 2025-02-09 #6966 @samhforbes
Hi @ManuelRausch  any joy with the above tasks?
