### 2025-02-12 #6966 @ManuelRausch
> Hi [@ManuelRausch](https://github.com/ManuelRausch) any joy with the above tasks?

I am still working on Latex Compatibility issues for the release on CRAN.
