+++ statConfR.Rcheck/00_pkg_src/statConfR/R/int_fitCEV.R
@@ -1,0 +1,103 @@
+###   Functions to fit the WEV model
+### Model version described by (Rausch et al., 2023)
+fitCEV <-
+  function(N_SA_RA, N_SA_RB, N_SB_RA, N_SB_RB,
+           nInits, nRestart, nRatings, nCond, nTrials){
+
+    # coarse grid search to find promising initial values
+
+    temp <- expand.grid(maxD =  seq(1, 5, 1),
+                        theta = seq(-1/2,1/2, 1/2),
+                        tauMin =  c(.1, .3, 1),  # position of the most conservative confidence criteria with respect to theta
+                        tauRange = seq(1, 5, 1),  #  position of the most liberal confidence criterion with respect to theta
+                        sigma = c(.1, .3, 1, 3), # noise parameter
+                        w = c(.1, .3, .6, .9)) # weighting parameter
+
+    inits <- data.frame(matrix(data=NA, nrow= nrow(temp), ncol = nCond + nRatings*2 + 1))
+    if(nCond==1)  {
+      inits[,1] <-  log(temp$maxD)  }
+    else{
+      inits[,1:(nCond)] <-  log(t(mapply(function(maxD) diff(seq(0, maxD, length.out = nCond+1)), temp$maxD)))
+    }
+    if (nRatings > 3){
+      inits[,(nCond+1):(nCond+nRatings-2)] <-
+        log(t(mapply(function(tauMin, tauRange) diff(seq(-tauRange-tauMin, -tauMin, length.out=nRatings-1)),
+                     temp$tauMin, temp$tauRange)))
+      inits[,(nCond+nRatings+2):(nCond + nRatings*2-1)] <-
+        log(t(mapply(function(tauMin, tauRange) diff(seq(tauMin, tauMin+tauRange, length.out=nRatings-1)),
+                     temp$tauMin, temp$tauRange)))
+    }
+    if (nRatings == 3){
+      inits[,(nCond+1):(nCond+nRatings-2)] <-
+        log(mapply(function(tauMin, tauRange) diff(seq(-tauRange-tauMin, -tauMin, length.out=nRatings-1)),
+                     temp$tauMin, temp$tauRange))
+      inits[,(nCond+nRatings+2):(nCond + nRatings*2-1)] <-
+        log(mapply(function(tauMin, tauRange) diff(seq(tauMin, tauMin+tauRange, length.out=nRatings-1)),
+                     temp$tauMin, temp$tauRange))
+    }
+    inits[,nCond+(nRatings-1)] <- -temp$tauMin
+    inits[,nCond+nRatings] <- temp$theta
+    inits[,nCond+(nRatings+1)] <- temp$tauMin
+    inits[,(nCond + nRatings*2)] <- log(temp$sigma)
+    inits[,(nCond + nRatings*2+1)] <- log(temp$w/(1-temp$w)) # logit transform w so it can vary between -Inf and Inf
+
+    logL <- apply(inits, MARGIN = 1,
+                  function(p) try(ll_CEV(p, N_SA_RA, N_SA_RB, N_SB_RA,N_SB_RB, nRatings, nCond), silent = TRUE))
+    logL <- as.numeric(logL)
+    inits <- inits[order(logL),]
+    inits <- inits[1:nInits,] #
+
+    # maximum likelihood optimization
+    noFitYet <- TRUE
+    for (i in 1:nInits){
+      m <- try(optim(par =  inits[i,],
+                     fn = ll_CEV, gr = NULL,
+                     N_SA_RA = N_SA_RA,N_SA_RB = N_SA_RB,
[…]
