+++ statConfR.Rcheck/00_pkg_src/statConfR/R/int_get_accuracy.R
@@ -1,0 +1,5 @@
+get_accuracy <- function(classifier)
+{
+  accuracy <- sum( apply(classifier, 2, max) )
+  accuracy
+}
