+++ statConfR.Rcheck/00_pkg_src/statConfR/R/int_get_pointwise_binary_transmitted_information.R
@@ -1,0 +1,4 @@
+get_pointwise_binary_transmitted_information <- pti2 <- function(c, p = 0.5)
+{
+  H2(p) - H2(c)
+}
