+++ statConfR.Rcheck/00_pkg_src/statConfR/R/int_get_binary_entropy.R
@@ -1,0 +1,5 @@
+get_binary_entropy <- H2 <- function(p) {
+  binary_entropy <- sapply(p, function(p) { get_entropy(c(p, 1-p)) } )
+  binary_entropy
+}
+
