+++ statConfR.Rcheck/00_pkg_src/statConfR/R/simulateConf.R
@@ -1,0 +1,106 @@
+#' @title Simulate data  according to a static model of confidence
+
+#' @description This function generates a data frame with random trials generated according to
+#' the computational model of decision confidence specified in the `model` argument
+#' with given parameters.
+#' Simulations can be used to visualize and test qualitative model predictions
+#' (e.g. using previously fitted parameters returned by \code{\link{fitConf}}).
+#' See \code{\link{fitConf}} for a full mathematical description of all models
+#' and their parameters.
+
+#' @param model `character` of length 1. The generative model that should be
+#'    used for simulation. Models implemented so far: 'WEV', 'SDT', 'GN', 'PDA',
+#'    'IG', 'ITGc', 'ITGcm', 'logN', and 'logWEV'.
+#' @param paramDf  a `data.frame`providing the number of generared trials and
+#' the parameters of the chosen model. `paramDf` should contain following columns
+#' (which parameters are needed depends on the specific model):
+#' * \code{N} (the number of trials be simulated),
+#' * \code{participant} (optional, the participant ID of each parameter set. Should be unique to each row),
+#' * \code{d_1}, \code{d_2}, ... (sensitivity parameters. The number of sensitivity parameters determines the number of levels of discriminability),
+#' * \code{c} (discrimination bias),
+#' * \code{theta_minus.1}, \code{theta_minus.2}, ... (confidence criteria associated with the response R = -1. The function simulates one more confidence category than there are confidence criteria),
+#' * \code{theta_plus.1}, \code{theta_plus.2}, ... (confidence criteria associated with the response R = 1. The function simulates one more confidence category than there are confidence criteria),
+#' * \code{w} (only for models WEV and logWEV: the visibility weighting parameter, bounded between 0 and 1),
+#' * \code{sigma} (only for models WEV, GN, logN, and logWEV: confidence noise, bounded between 0 and Inf),
+#' * \code{m} (only for IG, ITGm, and ITGcm: metacognitive efficiency parameter, bounded between 0 and Inf),
+#' * \code{b} (only for PDA: postdecisional accumulation parameter, bounded between 0 and Inf),
+#' * \code{M_theta_minus.1}, \code{M_theta_minus.2}, ... (only for logN: Mean confidence criteria associated with the response R = -1),
+#' * \code{M_theta_plus.1}, \code{M_theta_plus.2},... (only for logN: Mean confidence criteria associated with the response R = 1).
+
+#' @return a dataframe with about \code{nrow(paramDf)*N} rows (see Details),
+#' and the following columns:
+#' - \code{participant} giving the row ID of the simulation (see Details)
+#' - \code{stimulus} giving the category of the stimulus (-1 or 1)
+#' - only, if more than 1 sensitivity parameter (`d1`,`d2`,...) is provided:
+#' \code{diffCond} representing the difficulty condition (values correspond to
+#' the levels of the sensitivity parameters, i.e. diffCond=1 represents
+#' simulated trials with sensitivity `d1`)
+#' - \code{response} giving the response category (-1 or 1, corresponding to the stimulus categories)
+#' - \code{rating} giving the discrete confidence rating (integer, number of
+#' categories depends on the number of confidence criteria provided in the parameters)
+#' - \code{correct} giving the accuracy of the response (0 incorrect, 1 correct)
+#' - \code{ratings} same as `rating` but as a factor
+
+#' @details
+#' The function generates about `N` trials per row with the provided parameters
[…]
