### 2025-02-13 d0e9df9
Update plotConfModelFit.R
Make changes to pass CRAN checks
