### 2025-02-13 879994f
Adapting code to pass cran checks
