+++ man/plotConfModelFit.Rd
@@ -86,7 +86,7 @@ and overlays the predicted distribution over the data to which the model paramet
 # 2. Plot the predicted probabilities based on model and fitted parameters
   # against the observed relative frequencies.
 
-  - pa
+  PlotFitWEV <- plotConfModelFit(MaskOri, pars, model="WEV")
   PlotFitWEV
 
 }
