+++ statConfR.Rcheck/00_pkg_src/statConfR/DESCRIPTION
@@ -25,7 +25,7 @@ Date/Publication: 2023-09-12
 Encoding: UTF-8
 LazyData: true
 NeedsCompilation: no
-Packaged: 2025-02-12 22:03:04 UTC; PPA714
+Packaged: 2025-02-13 10:34:06 UTC; PPA714
 Repository: CRAN
 Roxygen: list(old_usage = TRUE, markdown=TRUE)
 RoxygenNote: 7.3.2
