+++ DESCRIPTION
@@ -20,7 +20,7 @@ License: GPL(>=3)
 URL: https://github.com/ManuelRausch/StatConfR
 BugReports: https://github.com/ManuelRausch/StatConfR/issues
 Depends:
-    R (>= 4.0)
+    R (>= 4.1.0)
 Imports:
     parallel, plyr, stats, utils, ggplot2, Rmisc
 Date/Publication: 2023-09-12
