+++ statConfR.Rcheck/00_pkg_src/statConfR/DESCRIPTION
@@ -19,13 +19,13 @@ Description: Provides fitting functions and other tools for decision confidence
 License: GPL(>=3)
 URL: https://github.com/ManuelRausch/StatConfR
 BugReports: https://github.com/ManuelRausch/StatConfR/issues
-Depends: R (>= 4.0)
+Depends: R (>= 4.1.0)
 Imports: parallel, plyr, stats, utils, ggplot2, Rmisc
 Date/Publication: 2023-09-12
 Encoding: UTF-8
 LazyData: true
 NeedsCompilation: no
-Packaged: 2025-02-13 13:05:16 UTC; PPA714
+Packaged: 2025-02-13 13:50:54 UTC; PPA714
 Repository: CRAN
 Roxygen: list(old_usage = TRUE, markdown=TRUE)
 RoxygenNote: 7.3.2
