### 2025-02-13 #6966 @ManuelRausch
- Finally, the package with all changes from the review has been released on CRAN with the version number 0.2.0.:https://cran.r-project.org/web/packages/statConfR/index.html
- The release has been archived on osf. doi:10.17605/OSF.IO/EQ3XK.
- We confirm that the authors and affiliations and ORCIDs are correct, both on the paper and on osf.
- The package on CRAN and on the archive have been released with the same licence (GPL-3).
