+++ paper.bib
@@ -1,190 +1,238 @@
 
-@article{desender_dynamic_2022,
-	title = {Dynamic influences on static measures of metacognition},
-	volume = {13},
-	doi = {10.1038/s41467-022-31727-0},
-	abstract = {Humans differ in their capability to judge the accuracy of their own choices via confidence judgments. Signal detection theory has been used to quantify the extent to which confidence tracks accuracy via M-ratio, often referred to as metacognitive efficiency. This measure, however, is static in that it does not consider the dynamics of decision making. This could be problematic because humans may shift their level of response caution to alter the tradeoff between speed and accuracy. Such shifts could induce unaccounted-for sources of variation in the assessment of metacognition. Instead, evidence accumulation frameworks consider decision making, including the computation of confidence, as a dynamic process unfolding over time. We draw on evidence accumulation frameworks to examine the influence of response caution on metacognition. Simulation results demonstrate that response caution has an influence on M-ratio. We then tested and confirmed that this was also the case in human participants who were explicitly instructed to either focus on speed or accuracy. We next demonstrated that this association between M-ratio and response caution was also present in an experiment without any reference towards speed. The latter finding was replicated in an independent dataset. In contrast, when data were analyzed with a novel dynamic measure of metacognition, which we refer to as v-ratio, in all of the three studies there was no effect of speed-accuracy tradeoff. These findings have important implications for research on metacognition, such as its measurement, domain-generality, individual differences, and neural correlates.Competing Interest StatementThe authors have declared no competing interest.},
+@article{daunizeau_vba_2014,
+	title = {{VBA}: A Probabilistic Treatment of Nonlinear Models for Neurobiological and Behavioural Data},
+	volume = {10},
+	issn = {1553-7358},
+	url = {https://dx.plos.org/10.1371/journal.pcbi.1003441},
+	doi = {10.1371/journal.pcbi.1003441},
+	shorttitle = {{VBA}},
+	abstract = {This work is in line with an on-going effort tending toward a computational (quantitative and refutable) understanding of human neuro-cognitive processes. Many sophisticated models for behavioural and neurobiological data have flourished during the past decade. Most of these models are partly unspecified (i.e. they have unknown parameters) and nonlinear. This makes them difficult to peer with a formal statistical data analysis framework. In turn, this compromises the reproducibility of model-based empirical studies. This work exposes a software toolbox that provides generic, efficient and robust probabilistic solutions to the three problems of model-based analysis of empirical data: (i) data simulation, (ii) parameter estimation/model selection, and (iii) experimental design optimization.},
+	pages = {e1003441},
 	number = {1},
-	journal = {Nature Communications},
-	author = {Desender, Kobe and Vermeylen, Luc and Verguts, Tom},
-	year = {2022},
-	pages = {1--30},
-	file = {PDF:C\:\\Users\\PPA714\\Zotero\\storage\\3WET758L\\Desender (2021) Dynamic influences on static measures of metacognition.pdf:application/pdf},
+	journaltitle = {{PLoS} Computational Biology},
+	shortjournal = {{PLoS} Comput Biol},
+	author = {Daunizeau, Jean and Adam, Vincent and Rigoux, Lionel},
+	editor = {Prlic, Andreas},
+	urldate = {2024-12-02},
+	date = {2014-01-23},
+	langid = {english},
[…]
