### 2025-02-21 #6966 @ManuelRausch
> Just had a read and all looks good, except in the references, where for journal articles the article title should be in sentence case (except after a colon), and the journal name should have first letters capitalised. See for example Hellman et al 2024 which is one that needs fixing.

Corrected.
