### 2025-02-21 #6966 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
